AG-16818-fix-preserve-group-order-and-filtering
