--- testing/behavioural/src/grouping-data/group-order-maintenance.test.ts
@@ -1,2521 +1,0 @@
-import { ClientSideRowModelModule, QuickFilterModule } from 'ag-grid-community';
-import { PivotModule, RowGroupingModule, TreeDataModule } from 'ag-grid-enterprise';
-
-import { GridColumns, GridRows, TestGridsManager, applyTransactionChecked } from '../test-utils';
-
-describe('group order maintenance', () => {
-    const gridsManager = new TestGridsManager({
-        modules: [QuickFilterModule, ClientSideRowModelModule, RowGroupingModule, PivotModule, TreeDataModule],
-    });
-
-    afterEach(() => gridsManager.reset());
-
-    test('per-level sort isolation: secondary sort cannot tie-break a non-targeted level', async () => {
-        const rowData = [
-            { id: '1', country: 'Alpha', sales: 10 },
-            { id: '2', country: 'Bravo', sales: 20 },
-            { id: '3', country: 'Charlie', sales: 30 },
-        ];
-
-        const api = gridsManager.createGrid('grid-isolation', {
-            columnDefs: [
-                { field: 'country', rowGroup: true, hide: true, sortable: true, comparator: () => 0 },
-                { field: 'sales', aggFunc: 'sum', sortable: true },
-            ],
-            autoGroupColumnDef: { headerName: 'Country' },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        api.applyColumnState({
-            state: [
-                { colId: 'country', sort: 'asc', sortIndex: 0 },
-                { colId: 'sales', sort: 'desc', sortIndex: 1 },
-            ],
-        });
-
-        await new GridRows(api, 'isolation: country in structural order').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Alpha ag-Grid-AutoColumn:"Alpha" sales:10
-            │ └── LEAF id:1 country:"Alpha" sales:10
-            ├─┬ LEAF_GROUP id:row-group-country-Bravo ag-Grid-AutoColumn:"Bravo" sales:20
-            │ └── LEAF id:2 country:"Bravo" sales:20
-            └─┬ LEAF_GROUP id:row-group-country-Charlie ag-Grid-AutoColumn:"Charlie" sales:30
-            · └── LEAF id:3 country:"Charlie" sales:30
-        `);
-    });
-
-    test('leaf-column sort with custom comparator inspecting aggData does not reorder groups under groupMaintainOrder', async () => {
-        const rowData = [
-            { id: '1', country: 'Italy', sales: 5 },
-            { id: '2', country: 'Italy', sales: 3 },
-            { id: '3', country: 'France', sales: 2 },
-            { id: '4', country: 'France', sales: 6 },
-            { id: '5', country: 'USA', sales: 100 },
-        ];
-
-        const api = gridsManager.createGrid('grid-aggdata-comparator', {
-            columnDefs: [
-                { field: 'country', rowGroup: true, hide: true },
-                {
-                    field: 'sales',
-                    aggFunc: 'sum',
-                    sortable: true,
-                    comparator: (_a, _b, nodeA, nodeB) => {
-                        const aggA = (nodeA as any)?.aggData?.sales ?? (nodeA as any)?.data?.sales ?? 0;
-                        const aggB = (nodeB as any)?.aggData?.sales ?? (nodeB as any)?.data?.sales ?? 0;
-                        return aggA - aggB;
-                    },
-                },
-            ],
-            autoGroupColumnDef: { headerName: 'Country' },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        api.applyColumnState({ state: [{ colId: 'sales', sort: 'desc' }] });
-
-        await new GridRows(api, 'aggData comparator: groups stay structural, leaves reorder').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy" sales:8
-            │ ├── LEAF id:1 country:"Italy" sales:5
-            │ └── LEAF id:2 country:"Italy" sales:3
-            ├─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France" sales:8
-            │ ├── LEAF id:4 country:"France" sales:6
-            │ └── LEAF id:3 country:"France" sales:2
-            └─┬ LEAF_GROUP id:row-group-country-USA ag-Grid-AutoColumn:"USA" sales:100
-            · └── LEAF id:5 country:"USA" sales:100
-        `);
-    });
-
-    test('uncoupled mode + aggFunc leaf-column sort: groups stay structural', async () => {
-        const rowData = [
-            { id: '1', country: 'Italy', sales: 5 },
-            { id: '2', country: 'Italy', sales: 3 },
-            { id: '3', country: 'France', sales: 2 },
-            { id: '4', country: 'France', sales: 100 },
-            { id: '5', country: 'USA', sales: 7 },
-        ];
-
-        const api = gridsManager.createGrid('grid-uncoupled-aggfunc-leaf', {
-            columnDefs: [
-                { field: 'country', rowGroup: true, hide: true },
-                { field: 'sales', aggFunc: 'sum', sortable: true },
-            ],
-            autoGroupColumnDef: {
-                headerName: 'Country',
-                comparator: (a: unknown, b: unknown) =>
-                    (a == null ? 0 : String(a).length) - (b == null ? 0 : String(b).length),
-            },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        api.applyColumnState({ state: [{ colId: 'sales', sort: 'desc' }] });
-
-        await new GridRows(api, 'uncoupled + aggFunc leaf sort: groups structural').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy" sales:8
-            │ ├── LEAF id:1 country:"Italy" sales:5
-            │ └── LEAF id:2 country:"Italy" sales:3
-            ├─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France" sales:102
-            │ ├── LEAF id:4 country:"France" sales:100
-            │ └── LEAF id:3 country:"France" sales:2
-            └─┬ LEAF_GROUP id:row-group-country-USA ag-Grid-AutoColumn:"USA" sales:7
-            · └── LEAF id:5 country:"USA" sales:7
-        `);
-    });
-
-    test('leaf rows are not reordered by a custom group-column comparator (data-row sort isolation)', async () => {
-        const rowData = [
-            { id: '1', country: 'Audi', athlete: 'Z' },
-            { id: '2', country: 'Audi', athlete: 'A' },
-            { id: '3', country: 'BMW', athlete: 'M' },
-        ];
-
-        const api = gridsManager.createGrid('grid-leaf-isolation', {
-            columnDefs: [
-                {
-                    field: 'country',
-                    rowGroup: true,
-                    hide: true,
-                    sortable: true,
-                    // Comparator that always returns 1 — even rows that share the group key would
-                    // get reordered if this option reached the data-row sort.
-                    comparator: () => 1,
-                },
-                { field: 'athlete' },
-            ],
-            autoGroupColumnDef: { headerName: 'Country' },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        api.applyColumnState({ state: [{ colId: 'country', sort: 'asc' }] });
-
-        await new GridRows(api, 'leaf isolation: data rows in data order').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            │ ├── LEAF id:1 country:"Audi" athlete:"Z"
-            │ └── LEAF id:2 country:"Audi" athlete:"A"
-            └─┬ LEAF_GROUP id:row-group-country-BMW ag-Grid-AutoColumn:"BMW"
-            · └── LEAF id:3 country:"BMW" athlete:"M"
-        `);
-    });
-
-    test('display column with own data: sort reaches both leaf rows AND group levels', async () => {
-        const rowData = [
-            { id: '1', country: 'Audi', athlete: 'A1', displayLabel: 'Z-display' },
-            { id: '2', country: 'Audi', athlete: 'A2', displayLabel: 'A-display' },
-            { id: '3', country: 'BMW', athlete: 'B1', displayLabel: 'M-display' },
-        ];
-
-        const api = gridsManager.createGrid('grid-display-data', {
-            columnDefs: [{ field: 'country', rowGroup: true, hide: true, comparator: () => 0 }, { field: 'athlete' }],
-            autoGroupColumnDef: {
-                headerName: 'Group',
-                showRowGroup: true,
-                field: 'displayLabel',
-                sortable: true,
-            },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        api.applyColumnState({ state: [{ colId: 'ag-Grid-AutoColumn', sort: 'asc' }] });
-
-        await new GridRows(api, 'display-data: leaf rows reordered by displayLabel').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            │ ├── LEAF id:2 ag-Grid-AutoColumn:"A-display" country:"Audi" athlete:"A2"
-            │ └── LEAF id:1 ag-Grid-AutoColumn:"Z-display" country:"Audi" athlete:"A1"
-            └─┬ LEAF_GROUP id:row-group-country-BMW ag-Grid-AutoColumn:"BMW"
-            · └── LEAF id:3 ag-Grid-AutoColumn:"M-display" country:"BMW" athlete:"B1"
-        `);
-    });
-
-    test('manual showRowGroup using source field (not colId) is not honoured by the grid — group order stays structural', async () => {
-        const rowData = [
-            { id: '1', country: 'Italy' },
-            { id: '2', country: 'France' },
-            { id: '3', country: 'Spain' },
-        ];
-
-        const api = gridsManager.createGrid('grid-manual-showrowgroup-by-field', {
-            columnDefs: [
-                {
-                    colId: 'customCountry',
-                    field: 'country',
-                    rowGroup: true,
-                    hide: true,
-                },
-                {
-                    colId: 'manualDisplay',
-                    headerName: 'Manual Display',
-                    showRowGroup: 'country', // field name, not colId — the grid does NOT resolve this
-                    sortable: true,
-                },
-            ],
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        const groupRowsBefore = api
-            .getRenderedNodes()
-            .filter((n) => n.group)
-            .map((n) => ({ key: n.key, groupData: n.groupData?.['manualDisplay'] }));
-        expect(groupRowsBefore.every((r) => r.groupData == null)).toBe(true);
-
-        api.applyColumnState({ state: [{ colId: 'manualDisplay', sort: 'desc' }] });
-        const groupOrderAfterSort = api
-            .getRenderedNodes()
-            .filter((n) => n.group)
-            .map((n) => n.key);
-        expect(groupOrderAfterSort).toEqual(['Italy', 'France', 'Spain']);
-    });
-
-    test('auto-display column with own field reorders groups under custom comparator (uncoupled escape hatch)', async () => {
-        const rowData = [
-            { id: '1', country: 'Audi', athlete: 'A1', displayLabel: 'long-label-Z' },
-            { id: '2', country: 'Audi', athlete: 'A2', displayLabel: 'short-A' },
-            { id: '3', country: 'BMW', athlete: 'B1', displayLabel: 'm-BMW' },
-            { id: '4', country: 'Tesla', athlete: 'T1', displayLabel: 'mid-T' },
-        ];
-
-        const api = gridsManager.createGrid('grid-display-regression', {
-            columnDefs: [{ field: 'country', rowGroup: true, hide: true }, { field: 'athlete' }],
-            autoGroupColumnDef: {
-                headerName: 'Group',
-                showRowGroup: true,
-                field: 'displayLabel',
-                sortable: true,
-                // Custom comparator on autoGroupColumnDef triggers uncoupled mode.
-                comparator: (a: unknown, b: unknown) => {
-                    const aLen = a == null ? 0 : String(a).length;
-                    const bLen = b == null ? 0 : String(b).length;
-                    return aLen - bLen;
-                },
-            },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        api.applyColumnState({ state: [{ colId: 'ag-Grid-AutoColumn', sort: 'asc' }] });
-
-        await new GridRows(api, 'auto-display sort: groups + leaves reorder by length').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-BMW ag-Grid-AutoColumn:"BMW"
-            │ └── LEAF id:3 ag-Grid-AutoColumn:"m-BMW" country:"BMW" athlete:"B1"
-            ├─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            │ ├── LEAF id:2 ag-Grid-AutoColumn:"short-A" country:"Audi" athlete:"A2"
-            │ └── LEAF id:1 ag-Grid-AutoColumn:"long-label-Z" country:"Audi" athlete:"A1"
-            └─┬ LEAF_GROUP id:row-group-country-Tesla ag-Grid-AutoColumn:"Tesla"
-            · └── LEAF id:4 ag-Grid-AutoColumn:"mid-T" country:"Tesla" athlete:"T1"
-        `);
-    });
-
-    test('display column with own comparator (no field/valueGetter): leaf rows sort by the comparator', async () => {
-        const rowData = [
-            { id: '1', country: 'Audi', athlete: 'Anna' },
-            { id: '2', country: 'Audi', athlete: 'Marco' },
-            { id: '3', country: 'BMW', athlete: 'Luca' },
-        ];
-
-        const api = gridsManager.createGrid('grid-comparator-only-display', {
-            columnDefs: [{ field: 'country', rowGroup: true, hide: true }, { field: 'athlete' }],
-            autoGroupColumnDef: {
-                headerName: 'Group',
-                showRowGroup: true,
-                sortable: true,
-                comparator: (_a, _b, nodeA, nodeB) => {
-                    const lenA = nodeA?.data?.athlete?.length ?? 0;
-                    const lenB = nodeB?.data?.athlete?.length ?? 0;
-                    return lenA - lenB;
-                },
-            },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        api.applyColumnState({ state: [{ colId: 'ag-Grid-AutoColumn', sort: 'desc' }] });
-
-        await new GridRows(api, 'comparator-only display column: leaves reorder by length desc').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            │ ├── LEAF id:2 country:"Audi" athlete:"Marco"
-            │ └── LEAF id:1 country:"Audi" athlete:"Anna"
-            └─┬ LEAF_GROUP id:row-group-country-BMW ag-Grid-AutoColumn:"BMW"
-            · └── LEAF id:3 country:"BMW" athlete:"Luca"
-        `);
-    });
-
-    test('empty leaf group: leaf-column sort is applied when data is later added by transaction', async () => {
-        const rowData = [
-            { id: '1', country: 'Audi', athlete: 'A1', total: 5 },
-            { id: '2', country: 'BMW', athlete: 'B1', total: 8 },
-        ];
-
-        const api = gridsManager.createGrid('grid-empty-leaf', {
-            columnDefs: [
-                { field: 'country', rowGroup: true, hide: true },
-                { field: 'athlete' },
-                { field: 'total', sortable: true },
-            ],
-            autoGroupColumnDef: { headerName: 'Country' },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        api.applyColumnState({ state: [{ colId: 'total', sort: 'desc' }] });
-
-        applyTransactionChecked(api, {
-            add: [
-                { id: '3', country: 'Tesla', athlete: 'T1', total: 4 },
-                { id: '4', country: 'Tesla', athlete: 'T2', total: 9 },
-            ],
-        });
-
-        await new GridRows(api, 'empty-leaf: new rows in leaf-sort order').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            │ └── LEAF id:1 country:"Audi" athlete:"A1" total:5
-            ├─┬ LEAF_GROUP id:row-group-country-BMW ag-Grid-AutoColumn:"BMW"
-            │ └── LEAF id:2 country:"BMW" athlete:"B1" total:8
-            └─┬ LEAF_GROUP id:row-group-country-Tesla ag-Grid-AutoColumn:"Tesla"
-            · ├── LEAF id:4 country:"Tesla" athlete:"T2" total:9
-            · └── LEAF id:3 country:"Tesla" athlete:"T1" total:4
-        `);
-    });
-
-    test('new group is appended at end when groupMaintainOrder is true', async () => {
-        const rowData = [
-            { id: '1', country: 'Ireland', athlete: 'I1' },
-            { id: '2', country: 'Ireland', athlete: 'I2' },
-            { id: '3', country: 'Italy', athlete: 'It1' },
-        ];
-
-        const api = gridsManager.createGrid('grid1', {
-            columnDefs: [{ field: 'country', rowGroup: true, hide: true }, { field: 'athlete' }],
-            autoGroupColumnDef: { headerName: 'Country' },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        await new GridRows(api, 'initial').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Ireland ag-Grid-AutoColumn:"Ireland"
-            │ ├── LEAF id:1 country:"Ireland" athlete:"I1"
-            │ └── LEAF id:2 country:"Ireland" athlete:"I2"
-            └─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            · └── LEAF id:3 country:"Italy" athlete:"It1"
-        `);
-
-        applyTransactionChecked(api, { add: [{ id: '4', country: 'France', athlete: 'F1' }] });
-
-        await new GridRows(api, 'after add France').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Ireland ag-Grid-AutoColumn:"Ireland"
-            │ ├── LEAF id:1 country:"Ireland" athlete:"I1"
-            │ └── LEAF id:2 country:"Ireland" athlete:"I2"
-            ├─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ └── LEAF id:3 country:"Italy" athlete:"It1"
-            └─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            · └── LEAF id:4 country:"France" athlete:"F1"
-        `);
-
-        await new GridColumns(api, 'columns').checkColumns(`
-            CENTER
-            ├── ag-Grid-AutoColumn "Country" width:200
-            └── athlete "Athlete" width:200
-        `);
-    });
-
-    test('updating a row without changing group does not change group order (groupMaintainOrder=true)', async () => {
-        const rowData = [
-            { id: '1', country: 'Ireland', athlete: 'I1' },
-            { id: '2', country: 'Ireland', athlete: 'I2' },
-            { id: '3', country: 'Italy', athlete: 'It1' },
-            { id: '4', country: 'France', athlete: 'F1' },
-        ];
-
-        const api = gridsManager.createGrid('grid2', {
-            columnDefs: [{ field: 'country', rowGroup: true, hide: true }, { field: 'athlete' }],
-            autoGroupColumnDef: { headerName: 'Country' },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        await new GridRows(api, 'initial').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Ireland ag-Grid-AutoColumn:"Ireland"
-            │ ├── LEAF id:1 country:"Ireland" athlete:"I1"
-            │ └── LEAF id:2 country:"Ireland" athlete:"I2"
-            ├─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ └── LEAF id:3 country:"Italy" athlete:"It1"
-            └─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            · └── LEAF id:4 country:"France" athlete:"F1"
-        `);
-
-        applyTransactionChecked(api, { update: [{ id: '2', country: 'Ireland', athlete: 'I2-upd' }] });
-
-        await new GridRows(api, 'after update').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Ireland ag-Grid-AutoColumn:"Ireland"
-            │ ├── LEAF id:1 country:"Ireland" athlete:"I1"
-            │ └── LEAF id:2 country:"Ireland" athlete:"I2-upd"
-            ├─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ └── LEAF id:3 country:"Italy" athlete:"It1"
-            └─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            · └── LEAF id:4 country:"France" athlete:"F1"
-        `);
-
-        await new GridColumns(api, 'columns').checkColumns(`
-            CENTER
-            ├── ag-Grid-AutoColumn "Country" width:200
-            └── athlete "Athlete" width:200
-        `);
-    });
-
-    test('group order is stable across rowData reorder (immutable mode, getRowId)', async () => {
-        let rowData = [
-            { id: '1', country: 'Audi', athlete: 'A' },
-            { id: '2', country: 'BMW', athlete: 'B' },
-            { id: '3', country: 'Tesla', athlete: 'T' },
-        ];
-
-        const api = gridsManager.createGrid('grid-reorder', {
-            columnDefs: [{ field: 'country', rowGroup: true, hide: true }, { field: 'athlete' }],
-            autoGroupColumnDef: { headerName: 'Country' },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        await new GridRows(api, 'reorder: initial').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            │ └── LEAF id:1 country:"Audi" athlete:"A"
-            ├─┬ LEAF_GROUP id:row-group-country-BMW ag-Grid-AutoColumn:"BMW"
-            │ └── LEAF id:2 country:"BMW" athlete:"B"
-            └─┬ LEAF_GROUP id:row-group-country-Tesla ag-Grid-AutoColumn:"Tesla"
-            · └── LEAF id:3 country:"Tesla" athlete:"T"
-        `);
-
-        rowData = [
-            { id: '3', country: 'Tesla', athlete: 'T' },
-            { id: '2', country: 'BMW', athlete: 'B' },
-            { id: '1', country: 'Audi', athlete: 'A' },
-        ];
-        api.setGridOption('rowData', rowData);
-
-        await new GridRows(api, 'reorder: groups stay in creation order').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            │ └── LEAF id:1 country:"Audi" athlete:"A"
-            ├─┬ LEAF_GROUP id:row-group-country-BMW ag-Grid-AutoColumn:"BMW"
-            │ └── LEAF id:2 country:"BMW" athlete:"B"
-            └─┬ LEAF_GROUP id:row-group-country-Tesla ag-Grid-AutoColumn:"Tesla"
-            · └── LEAF id:3 country:"Tesla" athlete:"T"
-        `);
-    });
-
-    test('postSortRows + groupMaintainOrder: customisation reapplies through filter cycles', async () => {
-        const rowData = [
-            { id: '1', country: 'Audi', athlete: 'A' },
-            { id: '2', country: 'BMW', athlete: 'B' },
-            { id: '3', country: 'Tesla', athlete: 'T' },
-        ];
-
-        const api = gridsManager.createGrid('grid-postsort', {
-            columnDefs: [{ field: 'country', rowGroup: true, hide: true }, { field: 'athlete' }],
-            autoGroupColumnDef: { headerName: 'Country' },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-            postSortRows: (params) => {
-                const teslaIdx = params.nodes.findIndex((n) => n.key === 'Tesla');
-                if (teslaIdx > 0) {
-                    const [tesla] = params.nodes.splice(teslaIdx, 1);
-                    params.nodes.unshift(tesla);
-                }
-            },
-        });
-
-        await new GridRows(api, 'postSort: initial — Tesla pinned to top').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Tesla ag-Grid-AutoColumn:"Tesla"
-            │ └── LEAF id:3 country:"Tesla" athlete:"T"
-            ├─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            │ └── LEAF id:1 country:"Audi" athlete:"A"
-            └─┬ LEAF_GROUP id:row-group-country-BMW ag-Grid-AutoColumn:"BMW"
-            · └── LEAF id:2 country:"BMW" athlete:"B"
-        `);
-
-        api.setGridOption('quickFilterText', 'BMW');
-        await new GridRows(api, 'postSort: filtered to BMW only').check(`
-            ROOT id:ROOT_NODE_ID
-            └─┬ LEAF_GROUP id:row-group-country-BMW ag-Grid-AutoColumn:"BMW"
-            · └── LEAF id:2 country:"BMW" athlete:"B"
-        `);
-
-        api.setGridOption('quickFilterText', undefined);
-        await new GridRows(api, 'postSort: clear filter — Tesla pinned again').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Tesla ag-Grid-AutoColumn:"Tesla"
-            │ └── LEAF id:3 country:"Tesla" athlete:"T"
-            ├─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            │ └── LEAF id:1 country:"Audi" athlete:"A"
-            └─┬ LEAF_GROUP id:row-group-country-BMW ag-Grid-AutoColumn:"BMW"
-            · └── LEAF id:2 country:"BMW" athlete:"B"
-        `);
-    });
-
-    test('grouped GroupSortStage: postSortRows reorder leaves childIndex / firstChild / lastChild stale (legacy AG-309 behaviour)', async () => {
-        // Locks the AG-309 (Feb 2018) ordering: _updateRowNodeAfterSort runs BEFORE postSortRows.
-        // The flags reflecting the pre-mutation order is intentional and out of scope to flip —
-        // changing it would break callers reading those flags during postSortRows.
-        const rowData = [
-            { id: '1', country: 'Audi', athlete: 'A' },
-            { id: '2', country: 'BMW', athlete: 'B' },
-            { id: '3', country: 'Tesla', athlete: 'T' },
-        ];
-
-        const api = gridsManager.createGrid('grid-group-postsort-stale', {
-            columnDefs: [{ field: 'country', rowGroup: true, hide: true }, { field: 'athlete' }],
-            autoGroupColumnDef: { headerName: 'Country' },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            rowData,
-            getRowId: (p) => p.data.id,
-            postSortRows: (params) => {
-                // Reverse the input. After this, childrenAfterSort is [Tesla, BMW, Audi] but
-                // the flags were already written for the input order [Audi, BMW, Tesla].
-                params.nodes.reverse();
-            },
-        });
-
-        // Display reflects the post-mutation array.
-        await new GridRows(api, 'AG-309 stale flags: displayed order is post-mutation').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Tesla ag-Grid-AutoColumn:"Tesla"
-            │ └── LEAF id:3 country:"Tesla" athlete:"T"
-            ├─┬ LEAF_GROUP id:row-group-country-BMW ag-Grid-AutoColumn:"BMW"
-            │ └── LEAF id:2 country:"BMW" athlete:"B"
-            └─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            · └── LEAF id:1 country:"Audi" athlete:"A"
-        `);
-
-        // Deprecated flags reflect the pre-mutation (input) order — Audi was first in input,
-        // Tesla was last, even though they're now at the opposite ends of the display.
-        const audi = api.getRowNode('row-group-country-Audi')!;
-        expect(audi.childIndex).toBe(0);
-        expect(audi.firstChild).toBe(true);
-        expect(audi.lastChild).toBe(false);
-
-        const tesla = api.getRowNode('row-group-country-Tesla')!;
-        expect(tesla.childIndex).toBe(2);
-        expect(tesla.firstChild).toBe(false);
-        expect(tesla.lastChild).toBe(true);
-    });
-
-    test('firstChild / lastChild / childIndex reflect the sorted group order', async () => {
-        const rowData = [
-            { id: '1', country: 'Audi', athlete: 'A' },
-            { id: '2', country: 'BMW', athlete: 'B' },
-            { id: '3', country: 'Tesla', athlete: 'T' },
-        ];
-
-        const api = gridsManager.createGrid('grid-group-sort-flags', {
-            columnDefs: [{ field: 'country', rowGroup: true, hide: true, sortable: true }, { field: 'athlete' }],
-            autoGroupColumnDef: { headerName: 'Country' },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        api.applyColumnState({ state: [{ colId: 'country', sort: 'desc' }] });
-
-        await new GridRows(api, 'group sort desc').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Tesla ag-Grid-AutoColumn:"Tesla"
-            │ └── LEAF id:3 country:"Tesla" athlete:"T"
-            ├─┬ LEAF_GROUP id:row-group-country-BMW ag-Grid-AutoColumn:"BMW"
-            │ └── LEAF id:2 country:"BMW" athlete:"B"
-            └─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            · └── LEAF id:1 country:"Audi" athlete:"A"
-        `);
-
-        const sortedKeys = ['Tesla', 'BMW', 'Audi'];
-        const groupNodes = sortedKeys.map((key) => api.getRowNode(`row-group-country-${key}`)!);
-
-        groupNodes.forEach((node, idx) => {
-            expect(node.childIndex).toBe(idx);
-            expect(node.firstChild).toBe(idx === 0);
-            expect(node.lastChild).toBe(idx === groupNodes.length - 1);
-        });
-    });
-
-    test('reused-array postSortRows mutation does not corrupt the structural baseline', async () => {
-        const rowData = [
-            { id: '1', country: 'Audi', athlete: 'A' },
-            { id: '2', country: 'BMW', athlete: 'B' },
-            { id: '3', country: 'Tesla', athlete: 'T' },
-        ];
-
-        let reverse = false;
-        const api = gridsManager.createGrid('grid-baseline-integrity', {
-            columnDefs: [{ field: 'country', rowGroup: true, hide: true }, { field: 'athlete' }],
-            autoGroupColumnDef: { headerName: 'Country' },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-            postSortRows: (params) => {
-                if (reverse) {
-                    params.nodes.reverse();
-                }
-            },
-        });
-
-        await new GridRows(api, 'baseline: structural order').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            │ └── LEAF id:1 country:"Audi" athlete:"A"
-            ├─┬ LEAF_GROUP id:row-group-country-BMW ag-Grid-AutoColumn:"BMW"
-            │ └── LEAF id:2 country:"BMW" athlete:"B"
-            └─┬ LEAF_GROUP id:row-group-country-Tesla ag-Grid-AutoColumn:"Tesla"
-            · └── LEAF id:3 country:"Tesla" athlete:"T"
-        `);
-
-        reverse = true;
-        api.refreshClientSideRowModel('sort');
-        await new GridRows(api, 'baseline: postSortRows reverses').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Tesla ag-Grid-AutoColumn:"Tesla"
-            │ └── LEAF id:3 country:"Tesla" athlete:"T"
-            ├─┬ LEAF_GROUP id:row-group-country-BMW ag-Grid-AutoColumn:"BMW"
-            │ └── LEAF id:2 country:"BMW" athlete:"B"
-            └─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            · └── LEAF id:1 country:"Audi" athlete:"A"
-        `);
-
-        reverse = false;
-        api.refreshClientSideRowModel('sort');
-        await new GridRows(api, 'baseline: structural order restored').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            │ └── LEAF id:1 country:"Audi" athlete:"A"
-            ├─┬ LEAF_GROUP id:row-group-country-BMW ag-Grid-AutoColumn:"BMW"
-            │ └── LEAF id:2 country:"BMW" athlete:"B"
-            └─┬ LEAF_GROUP id:row-group-country-Tesla ag-Grid-AutoColumn:"Tesla"
-            · └── LEAF id:3 country:"Tesla" athlete:"T"
-        `);
-    });
-
-    test('postSortRows reorder survives a sort refresh on the reused-array path', async () => {
-        const rowData = [
-            { id: '1', country: 'Audi', athlete: 'A' },
-            { id: '2', country: 'BMW', athlete: 'B' },
-            { id: '3', country: 'Tesla', athlete: 'T' },
-        ];
-
-        let promoteKey: string | null = null;
-        const api = gridsManager.createGrid('grid-postsort-reuse', {
-            columnDefs: [{ field: 'country', rowGroup: true, hide: true }, { field: 'athlete' }],
-            autoGroupColumnDef: { headerName: 'Country' },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-            postSortRows: (params) => {
-                if (!promoteKey) {
-                    return;
-                }
-                const idx = params.nodes.findIndex((n) => n.key === promoteKey);
-                if (idx > 0) {
-                    const [promoted] = params.nodes.splice(idx, 1);
-                    params.nodes.unshift(promoted);
-                }
-            },
-        });
-
-        await new GridRows(api, 'reuse-path: initial structural').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            │ └── LEAF id:1 country:"Audi" athlete:"A"
-            ├─┬ LEAF_GROUP id:row-group-country-BMW ag-Grid-AutoColumn:"BMW"
-            │ └── LEAF id:2 country:"BMW" athlete:"B"
-            └─┬ LEAF_GROUP id:row-group-country-Tesla ag-Grid-AutoColumn:"Tesla"
-            · └── LEAF id:3 country:"Tesla" athlete:"T"
-        `);
-
-        promoteKey = 'Tesla';
-        api.refreshClientSideRowModel('sort');
-        await new GridRows(api, 'reuse-path: Tesla promoted').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Tesla ag-Grid-AutoColumn:"Tesla"
-            │ └── LEAF id:3 country:"Tesla" athlete:"T"
-            ├─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            │ └── LEAF id:1 country:"Audi" athlete:"A"
-            └─┬ LEAF_GROUP id:row-group-country-BMW ag-Grid-AutoColumn:"BMW"
-            · └── LEAF id:2 country:"BMW" athlete:"B"
-        `);
-
-        api.refreshClientSideRowModel('sort');
-        await new GridRows(api, 'reuse-path: idempotent re-refresh').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Tesla ag-Grid-AutoColumn:"Tesla"
-            │ └── LEAF id:3 country:"Tesla" athlete:"T"
-            ├─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            │ └── LEAF id:1 country:"Audi" athlete:"A"
-            └─┬ LEAF_GROUP id:row-group-country-BMW ag-Grid-AutoColumn:"BMW"
-            · └── LEAF id:2 country:"BMW" athlete:"B"
-        `);
-
-        promoteKey = 'BMW';
-        api.refreshClientSideRowModel('sort');
-        await new GridRows(api, 'reuse-path: switch to BMW promoted').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-BMW ag-Grid-AutoColumn:"BMW"
-            │ └── LEAF id:2 country:"BMW" athlete:"B"
-            ├─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            │ └── LEAF id:1 country:"Audi" athlete:"A"
-            └─┬ LEAF_GROUP id:row-group-country-Tesla ag-Grid-AutoColumn:"Tesla"
-            · └── LEAF id:3 country:"Tesla" athlete:"T"
-        `);
-    });
-
-    test('groupHideOpenParents + postSortRows reorder on reused-array path: first-child change is detected', async () => {
-        const rowData = [
-            { id: '1', country: 'Audi', athlete: 'A1' },
-            { id: '2', country: 'Audi', athlete: 'A2' },
-            { id: '3', country: 'BMW', athlete: 'B1' },
-            { id: '4', country: 'BMW', athlete: 'B2' },
-        ];
-
-        let promoteId: string | null = null;
-        const api = gridsManager.createGrid('grid-ghop-reuse-firstchild', {
-            columnDefs: [{ field: 'country', rowGroup: true, hide: true }, { field: 'athlete' }],
-            autoGroupColumnDef: { headerName: 'Country', cellRendererParams: { suppressCount: true } },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            groupHideOpenParents: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-            postSortRows: (params) => {
-                if (!promoteId) {
-                    return;
-                }
-                const idx = params.nodes.findIndex((n) => n.id === promoteId);
-                if (idx > 0) {
-                    const [promoted] = params.nodes.splice(idx, 1);
-                    params.nodes.unshift(promoted);
-                }
-            },
-        });
-
-        await new GridRows(api, 'groupHideOpenParents-reuse: initial structural').check(`
-            ROOT id:ROOT_NODE_ID ag-Grid-AutoColumn-country:null
-            ├── LEAF id:1 ag-Grid-AutoColumn-country:"Audi" country:"Audi" athlete:"A1"
-            ├── LEAF id:2 country:"Audi" athlete:"A2"
-            ├── LEAF id:3 ag-Grid-AutoColumn-country:"BMW" country:"BMW" athlete:"B1"
-            └── LEAF id:4 country:"BMW" athlete:"B2"
-        `);
-
-        promoteId = '2';
-        api.refreshClientSideRowModel('sort');
-        await new GridRows(api, 'groupHideOpenParents-reuse: postSortRows promotes A2 inside Audi').check(`
-            ROOT id:ROOT_NODE_ID ag-Grid-AutoColumn-country:null
-            ├── LEAF id:2 ag-Grid-AutoColumn-country:"Audi" country:"Audi" athlete:"A2"
-            ├── LEAF id:1 country:"Audi" athlete:"A1"
-            ├── LEAF id:3 ag-Grid-AutoColumn-country:"BMW" country:"BMW" athlete:"B1"
-            └── LEAF id:4 country:"BMW" athlete:"B2"
-        `);
-
-        api.refreshClientSideRowModel('sort');
-        await new GridRows(api, 'groupHideOpenParents-reuse: idempotent re-refresh').check(`
-            ROOT id:ROOT_NODE_ID ag-Grid-AutoColumn-country:null
-            ├── LEAF id:2 ag-Grid-AutoColumn-country:"Audi" country:"Audi" athlete:"A2"
-            ├── LEAF id:1 country:"Audi" athlete:"A1"
-            ├── LEAF id:3 ag-Grid-AutoColumn-country:"BMW" country:"BMW" athlete:"B1"
-            └── LEAF id:4 country:"BMW" athlete:"B2"
-        `);
-
-        promoteId = '4';
-        api.refreshClientSideRowModel('sort');
-        await new GridRows(api, 'groupHideOpenParents-reuse: switch promote to B2 inside BMW').check(`
-            ROOT id:ROOT_NODE_ID ag-Grid-AutoColumn-country:null
-            ├── LEAF id:1 ag-Grid-AutoColumn-country:"Audi" country:"Audi" athlete:"A1"
-            ├── LEAF id:2 country:"Audi" athlete:"A2"
-            ├── LEAF id:4 ag-Grid-AutoColumn-country:"BMW" country:"BMW" athlete:"B2"
-            └── LEAF id:3 country:"BMW" athlete:"B1"
-        `);
-
-        promoteId = null;
-        api.refreshClientSideRowModel('sort');
-        await new GridRows(api, 'groupHideOpenParents-reuse: clear promotion — structural order restored').check(`
-            ROOT id:ROOT_NODE_ID ag-Grid-AutoColumn-country:null
-            ├── LEAF id:1 ag-Grid-AutoColumn-country:"Audi" country:"Audi" athlete:"A1"
-            ├── LEAF id:2 country:"Audi" athlete:"A2"
-            ├── LEAF id:3 ag-Grid-AutoColumn-country:"BMW" country:"BMW" athlete:"B1"
-            └── LEAF id:4 country:"BMW" athlete:"B2"
-        `);
-    });
-
-    test('runtime groupMaintainOrder toggle: option not in refreshProps, takes effect on next natural refresh', async () => {
-        const rowData = [
-            { id: '1', country: 'Italy', athlete: 'Z' },
-            { id: '2', country: 'France', athlete: 'A' },
-            { id: '3', country: 'Audi', athlete: 'M' },
-        ];
-
-        const api = gridsManager.createGrid('grid-runtime-toggle', {
-            columnDefs: [{ field: 'country', rowGroup: true, hide: true, sortable: true }, { field: 'athlete' }],
-            autoGroupColumnDef: { headerName: 'Country' },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: false,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        api.applyColumnState({ state: [{ colId: 'country', sort: 'desc' }] });
-        await new GridRows(api, 'toggle: country desc with maintainOrder=false').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ └── LEAF id:1 country:"Italy" athlete:"Z"
-            ├─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            │ └── LEAF id:2 country:"France" athlete:"A"
-            └─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            · └── LEAF id:3 country:"Audi" athlete:"M"
-        `);
-
-        api.setGridOption('groupMaintainOrder', true);
-        await new GridRows(api, 'toggle: ON without explicit refresh — previous sort persists').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ └── LEAF id:1 country:"Italy" athlete:"Z"
-            ├─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            │ └── LEAF id:2 country:"France" athlete:"A"
-            └─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            · └── LEAF id:3 country:"Audi" athlete:"M"
-        `);
-
-        api.refreshClientSideRowModel('sort');
-        await new GridRows(api, 'toggle: explicit refresh applies maintainOrder=true').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ └── LEAF id:1 country:"Italy" athlete:"Z"
-            ├─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            │ └── LEAF id:2 country:"France" athlete:"A"
-            └─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            · └── LEAF id:3 country:"Audi" athlete:"M"
-        `);
-
-        api.applyColumnState({ state: [{ colId: 'country', sort: null }] });
-        await new GridRows(api, 'toggle: clear country sort with maintainOrder=true').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ └── LEAF id:1 country:"Italy" athlete:"Z"
-            ├─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            │ └── LEAF id:2 country:"France" athlete:"A"
-            └─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            · └── LEAF id:3 country:"Audi" athlete:"M"
-        `);
-
-        api.setGridOption('groupMaintainOrder', false);
-        await new GridRows(api, 'toggle: OFF without explicit refresh — previous result persists').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ └── LEAF id:1 country:"Italy" athlete:"Z"
-            ├─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            │ └── LEAF id:2 country:"France" athlete:"A"
-            └─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            · └── LEAF id:3 country:"Audi" athlete:"M"
-        `);
-
-        api.applyColumnState({ state: [{ colId: 'country', sort: 'desc' }] });
-        await new GridRows(api, 'toggle: natural sort refresh applies maintainOrder=false').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ └── LEAF id:1 country:"Italy" athlete:"Z"
-            ├─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            │ └── LEAF id:2 country:"France" athlete:"A"
-            └─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            · └── LEAF id:3 country:"Audi" athlete:"M"
-        `);
-    });
-
-    test('pivot mode: leaf-group children are not reordered by an active leaf-column sort', async () => {
-        const rowData = [
-            { id: '1', country: 'Audi', year: 2020, athlete: 'Z' },
-            { id: '2', country: 'Audi', year: 2020, athlete: 'A' },
-            { id: '3', country: 'Audi', year: 2021, athlete: 'M' },
-        ];
-
-        const api = gridsManager.createGrid('grid-pivot-leaf-skip', {
-            columnDefs: [
-                { field: 'country', rowGroup: true, hide: true },
-                { field: 'year', pivot: true, hide: true },
-                { field: 'athlete', sort: 'asc' },
-            ],
-            autoGroupColumnDef: { headerName: 'Country' },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        const audi = () => api.getRowNode('row-group-country-Audi')!;
-        const childIds = () => audi().childrenAfterSort?.map((n) => n.id);
-
-        expect(childIds()).toEqual(['2', '3', '1']);
-
-        api.setGridOption('pivotMode', true);
-        expect(childIds()).toEqual(['1', '2', '3']);
-    });
-
-    test('pivot mode + groupMaintainOrder: filter cycle preserves group order', async () => {
-        const rowData = [
-            { id: '1', country: 'Audi', year: 2020, sales: 10 },
-            { id: '2', country: 'BMW', year: 2020, sales: 20 },
-            { id: '3', country: 'Tesla', year: 2021, sales: 30 },
-        ];
-
-        const api = gridsManager.createGrid('grid-pivot', {
-            columnDefs: [
-                { field: 'country', rowGroup: true, hide: true },
-                { field: 'year', pivot: true, hide: true },
-                { field: 'sales', aggFunc: 'sum' },
-            ],
-            autoGroupColumnDef: { headerName: 'Country' },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            pivotMode: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        const initialOrder = ['Audi', 'BMW', 'Tesla'];
-        const renderedKeys = () =>
-            api
-                .getRenderedNodes()
-                .filter((n) => n.level === 0 && n.group)
-                .map((n) => n.key);
-
-        expect(renderedKeys()).toEqual(initialOrder);
-
-        api.setGridOption('quickFilterText', 'BMW');
-        expect(renderedKeys()).toEqual(['BMW']);
-
-        api.setGridOption('quickFilterText', undefined);
-        expect(renderedKeys()).toEqual(initialOrder);
-    });
-
-    test('updating a row without changing group does not change group order (groupMaintainOrder=false)', async () => {
-        const rowData = [
-            { id: '1', country: 'Ireland', athlete: 'I1' },
-            { id: '2', country: 'Ireland', athlete: 'I2' },
-            { id: '3', country: 'Italy', athlete: 'It1' },
-            { id: '4', country: 'France', athlete: 'F1' },
-        ];
-
-        const api = gridsManager.createGrid('grid-update-no-group-change', {
-            columnDefs: [{ field: 'country', rowGroup: true, hide: true }, { field: 'athlete' }],
-            autoGroupColumnDef: { headerName: 'Country' },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: false,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        await new GridRows(api, 'initial').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Ireland ag-Grid-AutoColumn:"Ireland"
-            │ ├── LEAF id:1 country:"Ireland" athlete:"I1"
-            │ └── LEAF id:2 country:"Ireland" athlete:"I2"
-            ├─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ └── LEAF id:3 country:"Italy" athlete:"It1"
-            └─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            · └── LEAF id:4 country:"France" athlete:"F1"
-        `);
-
-        applyTransactionChecked(api, { update: [{ id: '2', country: 'Ireland', athlete: 'I2-upd' }] });
-
-        await new GridRows(api, 'after update').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Ireland ag-Grid-AutoColumn:"Ireland"
-            │ ├── LEAF id:1 country:"Ireland" athlete:"I1"
-            │ └── LEAF id:2 country:"Ireland" athlete:"I2-upd"
-            ├─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ └── LEAF id:3 country:"Italy" athlete:"It1"
-            └─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            · └── LEAF id:4 country:"France" athlete:"F1"
-        `);
-    });
-
-    test('leaf-column sort preserves group order (groupMaintainOrder=true)', async () => {
-        const rowData = [
-            { id: '1', country: 'Ireland', athlete: 'Zed' },
-            { id: '2', country: 'Italy', athlete: 'Ann' },
-            { id: '3', country: 'France', athlete: 'Mike' },
-        ];
-
-        const api = gridsManager.createGrid('grid-leaf-sort-preserves', {
-            columnDefs: [
-                { field: 'country', rowGroup: true, hide: true },
-                { field: 'athlete', sortable: true },
-            ],
-            autoGroupColumnDef: { headerName: 'Country' },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        await new GridRows(api, 'initial').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Ireland ag-Grid-AutoColumn:"Ireland"
-            │ └── LEAF id:1 country:"Ireland" athlete:"Zed"
-            ├─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ └── LEAF id:2 country:"Italy" athlete:"Ann"
-            └─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            · └── LEAF id:3 country:"France" athlete:"Mike"
-        `);
-
-        api.applyColumnState({ state: [{ colId: 'athlete', sort: 'asc' }] });
-
-        await new GridRows(api, 'leaf sort asc preserves group order').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Ireland ag-Grid-AutoColumn:"Ireland"
-            │ └── LEAF id:1 country:"Ireland" athlete:"Zed"
-            ├─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ └── LEAF id:2 country:"Italy" athlete:"Ann"
-            └─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            · └── LEAF id:3 country:"France" athlete:"Mike"
-        `);
-    });
-
-    test('group-column sort reorders groups (sorting coupled)', async () => {
-        const rowData = [
-            { id: '1', country: 'Ireland', athlete: 'A' },
-            { id: '2', country: 'Italy', athlete: 'B' },
-            { id: '3', country: 'France', athlete: 'C' },
-        ];
-
-        const api = gridsManager.createGrid('grid-group-col-sort-reorders', {
-            columnDefs: [
-                { field: 'country', rowGroup: true, hide: true, sortable: true },
-                { field: 'athlete', sortable: true },
-            ],
-            autoGroupColumnDef: { headerName: 'Country' },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        await new GridRows(api, 'initial').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Ireland ag-Grid-AutoColumn:"Ireland"
-            │ └── LEAF id:1 country:"Ireland" athlete:"A"
-            ├─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ └── LEAF id:2 country:"Italy" athlete:"B"
-            └─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            · └── LEAF id:3 country:"France" athlete:"C"
-        `);
-
-        api.applyColumnState({ state: [{ colId: 'country', sort: 'asc' }] });
-
-        await new GridRows(api, 'group sort asc').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            │ └── LEAF id:3 country:"France" athlete:"C"
-            ├─┬ LEAF_GROUP id:row-group-country-Ireland ag-Grid-AutoColumn:"Ireland"
-            │ └── LEAF id:1 country:"Ireland" athlete:"A"
-            └─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            · └── LEAF id:2 country:"Italy" athlete:"B"
-        `);
-
-        // Change to desc: Italy, Ireland, France
-        api.applyColumnState({ state: [{ colId: 'country', sort: 'desc' }] });
-
-        await new GridRows(api, 'group sort desc').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ └── LEAF id:2 country:"Italy" athlete:"B"
-            ├─┬ LEAF_GROUP id:row-group-country-Ireland ag-Grid-AutoColumn:"Ireland"
-            │ └── LEAF id:1 country:"Ireland" athlete:"A"
-            └─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            · └── LEAF id:3 country:"France" athlete:"C"
-        `);
-    });
-
-    test('adding a leaf sort while a group sort is active: each level sorts with its own options', async () => {
-        const rowData = [
-            { id: '1', country: 'Ireland', athlete: 'Z' },
-            { id: '2', country: 'Italy', athlete: 'A' },
-            { id: '3', country: 'France', athlete: 'M' },
-        ];
-
-        const api = gridsManager.createGrid('grid-leaf-sort-while-group-sort-active', {
-            columnDefs: [
-                { field: 'country', rowGroup: true, hide: true, sortable: true },
-                { field: 'athlete', sortable: true },
-            ],
-            autoGroupColumnDef: { headerName: 'Country' },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        await new GridRows(api, 'initial unsorted').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Ireland ag-Grid-AutoColumn:"Ireland"
-            │ └── LEAF id:1 country:"Ireland" athlete:"Z"
-            ├─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ └── LEAF id:2 country:"Italy" athlete:"A"
-            └─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            · └── LEAF id:3 country:"France" athlete:"M"
-        `);
-
-        api.applyColumnState({ state: [{ colId: 'country', sort: 'desc' }] });
-        await new GridRows(api, 'after group sort desc').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ └── LEAF id:2 country:"Italy" athlete:"A"
-            ├─┬ LEAF_GROUP id:row-group-country-Ireland ag-Grid-AutoColumn:"Ireland"
-            │ └── LEAF id:1 country:"Ireland" athlete:"Z"
-            └─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            · └── LEAF id:3 country:"France" athlete:"M"
-        `);
-
-        api.applyColumnState({ state: [{ colId: 'athlete', sort: 'asc' }] });
-        await new GridRows(api, 'leaf sort maintains last group order').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ └── LEAF id:2 country:"Italy" athlete:"A"
-            ├─┬ LEAF_GROUP id:row-group-country-Ireland ag-Grid-AutoColumn:"Ireland"
-            │ └── LEAF id:1 country:"Ireland" athlete:"Z"
-            └─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            · └── LEAF id:3 country:"France" athlete:"M"
-        `);
-    });
-
-    test('multi-level groupMaintainOrder: sort at one level only reorders that level', async () => {
-        const rowData = [
-            { id: '1', country: 'Italy', year: 2021, sales: 100 },
-            { id: '2', country: 'Italy', year: 2020, sales: 50 },
-            { id: '3', country: 'France', year: 2019, sales: 200 },
-            { id: '4', country: 'France', year: 2022, sales: 30 },
-            { id: '5', country: 'USA', year: 2018, sales: 70 },
-        ];
-
-        const api = gridsManager.createGrid('grid-multi-level-sort', {
-            columnDefs: [
-                { field: 'country', rowGroup: true, hide: true },
-                { field: 'year', rowGroup: true, hide: true },
-                { field: 'sales' },
-            ],
-            autoGroupColumnDef: { headerName: 'Group' },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        api.applyColumnState({ state: [{ colId: 'year', sort: 'asc' }] });
-        await new GridRows(api, 'inner (year) sort').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ filler id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ ├─┬ LEAF_GROUP id:row-group-country-Italy-year-2020 ag-Grid-AutoColumn:2020
-            │ │ └── LEAF id:2 country:"Italy" year:2020 sales:50
-            │ └─┬ LEAF_GROUP id:row-group-country-Italy-year-2021 ag-Grid-AutoColumn:2021
-            │ · └── LEAF id:1 country:"Italy" year:2021 sales:100
-            ├─┬ filler id:row-group-country-France ag-Grid-AutoColumn:"France"
-            │ ├─┬ LEAF_GROUP id:row-group-country-France-year-2019 ag-Grid-AutoColumn:2019
-            │ │ └── LEAF id:3 country:"France" year:2019 sales:200
-            │ └─┬ LEAF_GROUP id:row-group-country-France-year-2022 ag-Grid-AutoColumn:2022
-            │ · └── LEAF id:4 country:"France" year:2022 sales:30
-            └─┬ filler id:row-group-country-USA ag-Grid-AutoColumn:"USA"
-            · └─┬ LEAF_GROUP id:row-group-country-USA-year-2018 ag-Grid-AutoColumn:2018
-            · · └── LEAF id:5 country:"USA" year:2018 sales:70
-        `);
-
-        api.applyColumnState({
-            state: [{ colId: 'country', sort: 'asc' }],
-            defaultState: { sort: null },
-        });
-        await new GridRows(api, 'outer (country) sort').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ filler id:row-group-country-France ag-Grid-AutoColumn:"France"
-            │ ├─┬ LEAF_GROUP id:row-group-country-France-year-2019 ag-Grid-AutoColumn:2019
-            │ │ └── LEAF id:3 country:"France" year:2019 sales:200
-            │ └─┬ LEAF_GROUP id:row-group-country-France-year-2022 ag-Grid-AutoColumn:2022
-            │ · └── LEAF id:4 country:"France" year:2022 sales:30
-            ├─┬ filler id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ ├─┬ LEAF_GROUP id:row-group-country-Italy-year-2021 ag-Grid-AutoColumn:2021
-            │ │ └── LEAF id:1 country:"Italy" year:2021 sales:100
-            │ └─┬ LEAF_GROUP id:row-group-country-Italy-year-2020 ag-Grid-AutoColumn:2020
-            │ · └── LEAF id:2 country:"Italy" year:2020 sales:50
-            └─┬ filler id:row-group-country-USA ag-Grid-AutoColumn:"USA"
-            · └─┬ LEAF_GROUP id:row-group-country-USA-year-2018 ag-Grid-AutoColumn:2018
-            · · └── LEAF id:5 country:"USA" year:2018 sales:70
-        `);
-    });
-
-    test('multi-level groupMaintainOrder + secondary leaf sort: leaf rows sort inside each leaf group', async () => {
-        const rowData = [
-            { id: '1', country: 'Italy', year: 2021, athlete: 'Zed' },
-            { id: '2', country: 'Italy', year: 2021, athlete: 'Anna' },
-            { id: '3', country: 'France', year: 2019, athlete: 'Mark' },
-            { id: '4', country: 'France', year: 2019, athlete: 'Bob' },
-        ];
-
-        const api = gridsManager.createGrid('grid-leaf-sort-secondary', {
-            columnDefs: [
-                { field: 'country', rowGroup: true, hide: true },
-                { field: 'year', rowGroup: true, hide: true },
-                { field: 'athlete' },
-            ],
-            autoGroupColumnDef: { headerName: 'Group' },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        api.applyColumnState({
-            state: [
-                { colId: 'country', sort: 'asc', sortIndex: 0 },
-                { colId: 'athlete', sort: 'asc', sortIndex: 1 },
-            ],
-        });
-
-        await new GridRows(api, 'leaf-sort-secondary: country reorders, leaf rows sort by athlete').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ filler id:row-group-country-France ag-Grid-AutoColumn:"France"
-            │ └─┬ LEAF_GROUP id:row-group-country-France-year-2019 ag-Grid-AutoColumn:2019
-            │ · ├── LEAF id:4 country:"France" year:2019 athlete:"Bob"
-            │ · └── LEAF id:3 country:"France" year:2019 athlete:"Mark"
-            └─┬ filler id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            · └─┬ LEAF_GROUP id:row-group-country-Italy-year-2021 ag-Grid-AutoColumn:2021
-            · · ├── LEAF id:2 country:"Italy" year:2021 athlete:"Anna"
-            · · └── LEAF id:1 country:"Italy" year:2021 athlete:"Zed"
-        `);
-    });
-
-    test('singleColumn + groupMaintainOrder: cascade-equivalent group sort reorders every level', async () => {
-        const rowData = [
-            { id: '1', country: 'Italy', year: 2021, sales: 100 },
-            { id: '2', country: 'Italy', year: 2020, sales: 50 },
-            { id: '3', country: 'France', year: 2019, sales: 200 },
-            { id: '4', country: 'France', year: 2022, sales: 30 },
-        ];
-
-        const api = gridsManager.createGrid('grid-single-col-cascade', {
-            columnDefs: [
-                { field: 'country', rowGroup: true, hide: true },
-                { field: 'year', rowGroup: true, hide: true },
-                { field: 'sales' },
-            ],
-            autoGroupColumnDef: { headerName: 'Group' },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        api.applyColumnState({
-            state: [
-                { colId: 'country', sort: 'asc', sortIndex: 0 },
-                { colId: 'year', sort: 'asc', sortIndex: 1 },
-            ],
-        });
-
-        await new GridRows(api, 'singleColumn cascade: country and year both reorder').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ filler id:row-group-country-France ag-Grid-AutoColumn:"France"
-            │ ├─┬ LEAF_GROUP id:row-group-country-France-year-2019 ag-Grid-AutoColumn:2019
-            │ │ └── LEAF id:3 country:"France" year:2019 sales:200
-            │ └─┬ LEAF_GROUP id:row-group-country-France-year-2022 ag-Grid-AutoColumn:2022
-            │ · └── LEAF id:4 country:"France" year:2022 sales:30
-            └─┬ filler id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            · ├─┬ LEAF_GROUP id:row-group-country-Italy-year-2020 ag-Grid-AutoColumn:2020
-            · │ └── LEAF id:2 country:"Italy" year:2020 sales:50
-            · └─┬ LEAF_GROUP id:row-group-country-Italy-year-2021 ag-Grid-AutoColumn:2021
-            · · └── LEAF id:1 country:"Italy" year:2021 sales:100
-        `);
-    });
-
-    test('singleColumn cascade with 5 group levels: every level reorders', async () => {
-        const rowData = [
-            { id: '1', a: 'Z', b: 'X', c: 'Q', d: 'M', e: 'C' },
-            { id: '2', a: 'A', b: 'Y', c: 'P', d: 'L', e: 'B' },
-            { id: '3', a: 'M', b: 'B', c: 'O', d: 'K', e: 'A' },
-            { id: '4', a: 'C', b: 'A', c: 'N', d: 'J', e: 'D' },
-        ];
-
-        const api = gridsManager.createGrid('grid-five-level-cascade', {
-            columnDefs: [
-                { field: 'a', rowGroup: true, hide: true },
-                { field: 'b', rowGroup: true, hide: true },
-                { field: 'c', rowGroup: true, hide: true },
-                { field: 'd', rowGroup: true, hide: true },
-                { field: 'e', rowGroup: true, hide: true },
-            ],
-            autoGroupColumnDef: { headerName: 'Group' },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        await new GridColumns(api, 'five-level: single auto-display column').checkColumns(`
-            CENTER
-            └── ag-Grid-AutoColumn "Group" width:200
-        `);
-
-        api.applyColumnState({
-            state: [
-                { colId: 'a', sort: 'asc', sortIndex: 0 },
-                { colId: 'b', sort: 'asc', sortIndex: 1 },
-                { colId: 'c', sort: 'asc', sortIndex: 2 },
-                { colId: 'd', sort: 'asc', sortIndex: 3 },
-                { colId: 'e', sort: 'asc', sortIndex: 4 },
-            ],
-        });
-
-        await new GridRows(api, 'five-level cascade asc: every level reorders').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ filler id:row-group-a-A ag-Grid-AutoColumn:"A"
-            │ └─┬ filler id:row-group-a-A-b-Y ag-Grid-AutoColumn:"Y"
-            │ · └─┬ filler id:row-group-a-A-b-Y-c-P ag-Grid-AutoColumn:"P"
-            │ · · └─┬ filler id:row-group-a-A-b-Y-c-P-d-L ag-Grid-AutoColumn:"L"
-            │ · · · └─┬ LEAF_GROUP id:row-group-a-A-b-Y-c-P-d-L-e-B ag-Grid-AutoColumn:"B"
-            │ · · · · └── LEAF id:2 a:"A" b:"Y" c:"P" d:"L" e:"B"
-            ├─┬ filler id:row-group-a-C ag-Grid-AutoColumn:"C"
-            │ └─┬ filler id:row-group-a-C-b-A ag-Grid-AutoColumn:"A"
-            │ · └─┬ filler id:row-group-a-C-b-A-c-N ag-Grid-AutoColumn:"N"
-            │ · · └─┬ filler id:row-group-a-C-b-A-c-N-d-J ag-Grid-AutoColumn:"J"
-            │ · · · └─┬ LEAF_GROUP id:row-group-a-C-b-A-c-N-d-J-e-D ag-Grid-AutoColumn:"D"
-            │ · · · · └── LEAF id:4 a:"C" b:"A" c:"N" d:"J" e:"D"
-            ├─┬ filler id:row-group-a-M ag-Grid-AutoColumn:"M"
-            │ └─┬ filler id:row-group-a-M-b-B ag-Grid-AutoColumn:"B"
-            │ · └─┬ filler id:row-group-a-M-b-B-c-O ag-Grid-AutoColumn:"O"
-            │ · · └─┬ filler id:row-group-a-M-b-B-c-O-d-K ag-Grid-AutoColumn:"K"
-            │ · · · └─┬ LEAF_GROUP id:row-group-a-M-b-B-c-O-d-K-e-A ag-Grid-AutoColumn:"A"
-            │ · · · · └── LEAF id:3 a:"M" b:"B" c:"O" d:"K" e:"A"
-            └─┬ filler id:row-group-a-Z ag-Grid-AutoColumn:"Z"
-            · └─┬ filler id:row-group-a-Z-b-X ag-Grid-AutoColumn:"X"
-            · · └─┬ filler id:row-group-a-Z-b-X-c-Q ag-Grid-AutoColumn:"Q"
-            · · · └─┬ filler id:row-group-a-Z-b-X-c-Q-d-M ag-Grid-AutoColumn:"M"
-            · · · · └─┬ LEAF_GROUP id:row-group-a-Z-b-X-c-Q-d-M-e-C ag-Grid-AutoColumn:"C"
-            · · · · · └── LEAF id:1 a:"Z" b:"X" c:"Q" d:"M" e:"C"
-        `);
-
-        // Flip to descending — order reverses on every level: Z → M → C → A (ids 1,3,4,2).
-        api.applyColumnState({
-            state: [
-                { colId: 'a', sort: 'desc', sortIndex: 0 },
-                { colId: 'b', sort: 'desc', sortIndex: 1 },
-                { colId: 'c', sort: 'desc', sortIndex: 2 },
-                { colId: 'd', sort: 'desc', sortIndex: 3 },
-                { colId: 'e', sort: 'desc', sortIndex: 4 },
-            ],
-        });
-
-        await new GridRows(api, 'five-level cascade desc: every level reverses').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ filler id:row-group-a-Z ag-Grid-AutoColumn:"Z"
-            │ └─┬ filler id:row-group-a-Z-b-X ag-Grid-AutoColumn:"X"
-            │ · └─┬ filler id:row-group-a-Z-b-X-c-Q ag-Grid-AutoColumn:"Q"
-            │ · · └─┬ filler id:row-group-a-Z-b-X-c-Q-d-M ag-Grid-AutoColumn:"M"
-            │ · · · └─┬ LEAF_GROUP id:row-group-a-Z-b-X-c-Q-d-M-e-C ag-Grid-AutoColumn:"C"
-            │ · · · · └── LEAF id:1 a:"Z" b:"X" c:"Q" d:"M" e:"C"
-            ├─┬ filler id:row-group-a-M ag-Grid-AutoColumn:"M"
-            │ └─┬ filler id:row-group-a-M-b-B ag-Grid-AutoColumn:"B"
-            │ · └─┬ filler id:row-group-a-M-b-B-c-O ag-Grid-AutoColumn:"O"
-            │ · · └─┬ filler id:row-group-a-M-b-B-c-O-d-K ag-Grid-AutoColumn:"K"
-            │ · · · └─┬ LEAF_GROUP id:row-group-a-M-b-B-c-O-d-K-e-A ag-Grid-AutoColumn:"A"
-            │ · · · · └── LEAF id:3 a:"M" b:"B" c:"O" d:"K" e:"A"
-            ├─┬ filler id:row-group-a-C ag-Grid-AutoColumn:"C"
-            │ └─┬ filler id:row-group-a-C-b-A ag-Grid-AutoColumn:"A"
-            │ · └─┬ filler id:row-group-a-C-b-A-c-N ag-Grid-AutoColumn:"N"
-            │ · · └─┬ filler id:row-group-a-C-b-A-c-N-d-J ag-Grid-AutoColumn:"J"
-            │ · · · └─┬ LEAF_GROUP id:row-group-a-C-b-A-c-N-d-J-e-D ag-Grid-AutoColumn:"D"
-            │ · · · · └── LEAF id:4 a:"C" b:"A" c:"N" d:"J" e:"D"
-            └─┬ filler id:row-group-a-A ag-Grid-AutoColumn:"A"
-            · └─┬ filler id:row-group-a-A-b-Y ag-Grid-AutoColumn:"Y"
-            · · └─┬ filler id:row-group-a-A-b-Y-c-P ag-Grid-AutoColumn:"P"
-            · · · └─┬ filler id:row-group-a-A-b-Y-c-P-d-L ag-Grid-AutoColumn:"L"
-            · · · · └─┬ LEAF_GROUP id:row-group-a-A-b-Y-c-P-d-L-e-B ag-Grid-AutoColumn:"B"
-            · · · · · └── LEAF id:2 a:"A" b:"Y" c:"P" d:"L" e:"B"
-        `);
-    });
-
-    test('multipleColumns + custom colId on rowGroup column: per-level isolation is preserved', async () => {
-        // Each rowGroup column has a custom colId; multipleColumns generates one auto-display
-        // column per level whose `showRowGroup` is the source colId. Locks in the unified
-        // buildLevelSortTargeted lookup: matches source rowGroup columns by reference AND
-        // auto-display columns by `colDef.showRowGroup`, regardless of coupling.
-        const rowData = [
-            { id: '1', country: 'Italy', year: 2021, sales: 100 },
-            { id: '2', country: 'Italy', year: 2020, sales: 50 },
-            { id: '3', country: 'France', year: 2019, sales: 200 },
-            { id: '4', country: 'France', year: 2022, sales: 30 },
-        ];
-
-        const api = gridsManager.createGrid('grid-multi-cols-custom-colid', {
-            columnDefs: [
-                { colId: 'customCountry', field: 'country', rowGroup: true, hide: true },
-                { colId: 'customYear', field: 'year', rowGroup: true, hide: true },
-                { field: 'sales' },
-            ],
-            groupDisplayType: 'multipleColumns',
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        api.applyColumnState({ state: [{ colId: 'customYear', sort: 'asc' }] });
-
-        // Country groups stay in structural order (Italy, France); year groups within each
-        // country are sorted ascending.
-        await new GridRows(api, 'multipleColumns: year sort respects per-level isolation').check(`
-            ROOT id:ROOT_NODE_ID ag-Grid-AutoColumn-customCountry:null ag-Grid-AutoColumn-customYear:null
-            ├─┬ filler id:row-group-customCountry-Italy ag-Grid-AutoColumn-customCountry:"Italy" ag-Grid-AutoColumn-customYear:null
-            │ ├─┬ LEAF_GROUP id:row-group-customCountry-Italy-customYear-2020 ag-Grid-AutoColumn-customYear:2020
-            │ │ └── LEAF id:2 customCountry:"Italy" customYear:2020 sales:50
-            │ └─┬ LEAF_GROUP id:row-group-customCountry-Italy-customYear-2021 ag-Grid-AutoColumn-customYear:2021
-            │ · └── LEAF id:1 customCountry:"Italy" customYear:2021 sales:100
-            └─┬ filler id:row-group-customCountry-France ag-Grid-AutoColumn-customCountry:"France" ag-Grid-AutoColumn-customYear:null
-            · ├─┬ LEAF_GROUP id:row-group-customCountry-France-customYear-2019 ag-Grid-AutoColumn-customYear:2019
-            · │ └── LEAF id:3 customCountry:"France" customYear:2019 sales:200
-            · └─┬ LEAF_GROUP id:row-group-customCountry-France-customYear-2022 ag-Grid-AutoColumn-customYear:2022
-            · · └── LEAF id:4 customCountry:"France" customYear:2022 sales:30
-        `);
-
-        await new GridColumns(api, 'multipleColumns: auto-display columns + sales').checkColumns(`
-            CENTER
-            ├── ag-Grid-AutoColumn-customCountry "Country" width:200
-            ├── ag-Grid-AutoColumn-customYear "Year" width:200
-            └── sales "Sales" width:200
-        `);
-    });
-
-    test('clearing a filter restores the original group order', async () => {
-        const rowData = [
-            { id: '1', country: 'Ireland', athlete: 'I1' },
-            { id: '2', country: 'Italy', athlete: 'T1' },
-            { id: '3', country: 'France', athlete: 'F1' },
-            { id: '4', country: 'Spain', athlete: 'S1' },
-        ];
-
-        const api = gridsManager.createGrid('grid-filter-reset', {
-            columnDefs: [{ field: 'country', rowGroup: true, hide: true }, { field: 'athlete' }],
-            autoGroupColumnDef: { headerName: 'Country' },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        api.setGridOption('quickFilterText', 'T1');
-        await new GridRows(api, 'after filter to Italy only').check(`
-            ROOT id:ROOT_NODE_ID
-            └─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            · └── LEAF id:2 country:"Italy" athlete:"T1"
-        `);
-
-        api.setGridOption('quickFilterText', undefined);
-        await new GridRows(api, 'after clearing filter — original order').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Ireland ag-Grid-AutoColumn:"Ireland"
-            │ └── LEAF id:1 country:"Ireland" athlete:"I1"
-            ├─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ └── LEAF id:2 country:"Italy" athlete:"T1"
-            ├─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            │ └── LEAF id:3 country:"France" athlete:"F1"
-            └─┬ LEAF_GROUP id:row-group-country-Spain ag-Grid-AutoColumn:"Spain"
-            · └── LEAF id:4 country:"Spain" athlete:"S1"
-        `);
-    });
-
-    test('filter cycle interleaved with add / update / remove transactions', async () => {
-        const rowData = [
-            { id: '1', country: 'Audi', athlete: 'Anna' },
-            { id: '2', country: 'BMW', athlete: 'Bert' },
-            { id: '3', country: 'Tesla', athlete: 'Tim' },
-        ];
-
-        const api = gridsManager.createGrid('grid-stress', {
-            columnDefs: [{ field: 'country', rowGroup: true, hide: true }, { field: 'athlete' }],
-            autoGroupColumnDef: { headerName: 'Country' },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        await new GridRows(api, 'stress: initial').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            │ └── LEAF id:1 country:"Audi" athlete:"Anna"
-            ├─┬ LEAF_GROUP id:row-group-country-BMW ag-Grid-AutoColumn:"BMW"
-            │ └── LEAF id:2 country:"BMW" athlete:"Bert"
-            └─┬ LEAF_GROUP id:row-group-country-Tesla ag-Grid-AutoColumn:"Tesla"
-            · └── LEAF id:3 country:"Tesla" athlete:"Tim"
-        `);
-
-        api.setGridOption('quickFilterText', 'Tim');
-        await new GridRows(api, 'stress: filter Tesla').check(`
-            ROOT id:ROOT_NODE_ID
-            └─┬ LEAF_GROUP id:row-group-country-Tesla ag-Grid-AutoColumn:"Tesla"
-            · └── LEAF id:3 country:"Tesla" athlete:"Tim"
-        `);
-
-        applyTransactionChecked(api, { add: [{ id: '4', country: 'Volvo', athlete: 'Timmy' }] });
-        await new GridRows(api, 'stress: add Volvo (visible) while filtered').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Tesla ag-Grid-AutoColumn:"Tesla"
-            │ └── LEAF id:3 country:"Tesla" athlete:"Tim"
-            └─┬ LEAF_GROUP id:row-group-country-Volvo ag-Grid-AutoColumn:"Volvo"
-            · └── LEAF id:4 country:"Volvo" athlete:"Timmy"
-        `);
-
-        applyTransactionChecked(api, { update: [{ id: '1', country: 'Audi', athlete: 'Anna-upd' }] });
-        await new GridRows(api, 'stress: update hidden Audi while filtered').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Tesla ag-Grid-AutoColumn:"Tesla"
-            │ └── LEAF id:3 country:"Tesla" athlete:"Tim"
-            └─┬ LEAF_GROUP id:row-group-country-Volvo ag-Grid-AutoColumn:"Volvo"
-            · └── LEAF id:4 country:"Volvo" athlete:"Timmy"
-        `);
-
-        applyTransactionChecked(api, { remove: [{ id: '2', country: 'BMW', athlete: 'Bert' }] });
-        await new GridRows(api, 'stress: remove hidden BMW while filtered').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Tesla ag-Grid-AutoColumn:"Tesla"
-            │ └── LEAF id:3 country:"Tesla" athlete:"Tim"
-            └─┬ LEAF_GROUP id:row-group-country-Volvo ag-Grid-AutoColumn:"Volvo"
-            · └── LEAF id:4 country:"Volvo" athlete:"Timmy"
-        `);
-
-        api.setGridOption('quickFilterText', undefined);
-        await new GridRows(api, 'stress: clear filter — original positions restored').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            │ └── LEAF id:1 country:"Audi" athlete:"Anna-upd"
-            ├─┬ LEAF_GROUP id:row-group-country-Tesla ag-Grid-AutoColumn:"Tesla"
-            │ └── LEAF id:3 country:"Tesla" athlete:"Tim"
-            └─┬ LEAF_GROUP id:row-group-country-Volvo ag-Grid-AutoColumn:"Volvo"
-            · └── LEAF id:4 country:"Volvo" athlete:"Timmy"
-        `);
-    });
-
-    test('initialGroupOrderComparator + groupMaintainOrder + filter cycle', async () => {
-        const rowData = [
-            { id: '1', country: 'Tesla', athlete: 'Tim' },
-            { id: '2', country: 'BMW', athlete: 'Bert' },
-            { id: '3', country: 'Audi', athlete: 'Anna' },
-        ];
-
-        const api = gridsManager.createGrid('grid-comparator', {
-            columnDefs: [{ field: 'country', rowGroup: true, hide: true }, { field: 'athlete' }],
-            autoGroupColumnDef: { headerName: 'Country' },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            initialGroupOrderComparator: ({ nodeA, nodeB }) => (nodeA.key! < nodeB.key! ? -1 : 1),
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        await new GridRows(api, 'comparator: initial alphabetical order').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            │ └── LEAF id:3 country:"Audi" athlete:"Anna"
-            ├─┬ LEAF_GROUP id:row-group-country-BMW ag-Grid-AutoColumn:"BMW"
-            │ └── LEAF id:2 country:"BMW" athlete:"Bert"
-            └─┬ LEAF_GROUP id:row-group-country-Tesla ag-Grid-AutoColumn:"Tesla"
-            · └── LEAF id:1 country:"Tesla" athlete:"Tim"
-        `);
-
-        api.setGridOption('quickFilterText', 'Tim');
-        await new GridRows(api, 'comparator: filter Tesla only').check(`
-            ROOT id:ROOT_NODE_ID
-            └─┬ LEAF_GROUP id:row-group-country-Tesla ag-Grid-AutoColumn:"Tesla"
-            · └── LEAF id:1 country:"Tesla" athlete:"Tim"
-        `);
-
-        api.setGridOption('quickFilterText', undefined);
-        await new GridRows(api, 'comparator: clear filter — alphabetical order restored').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            │ └── LEAF id:3 country:"Audi" athlete:"Anna"
-            ├─┬ LEAF_GROUP id:row-group-country-BMW ag-Grid-AutoColumn:"BMW"
-            │ └── LEAF id:2 country:"BMW" athlete:"Bert"
-            └─┬ LEAF_GROUP id:row-group-country-Tesla ag-Grid-AutoColumn:"Tesla"
-            · └── LEAF id:1 country:"Tesla" athlete:"Tim"
-        `);
-
-        applyTransactionChecked(api, { add: [{ id: '4', country: 'Acura', athlete: 'Alex' }] });
-        await new GridRows(api, 'comparator: add Acura via transaction — sorted alphabetically').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Acura ag-Grid-AutoColumn:"Acura"
-            │ └── LEAF id:4 country:"Acura" athlete:"Alex"
-            ├─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            │ └── LEAF id:3 country:"Audi" athlete:"Anna"
-            ├─┬ LEAF_GROUP id:row-group-country-BMW ag-Grid-AutoColumn:"BMW"
-            │ └── LEAF id:2 country:"BMW" athlete:"Bert"
-            └─┬ LEAF_GROUP id:row-group-country-Tesla ag-Grid-AutoColumn:"Tesla"
-            · └── LEAF id:1 country:"Tesla" athlete:"Tim"
-        `);
-    });
-
-    test('after filtering removes a group, adding a new group appends at end', async () => {
-        const rowData = [
-            { id: '1', country: 'Ireland', athlete: 'I1' },
-            { id: '2', country: 'Italy', athlete: 'T1' },
-            { id: '3', country: 'France', athlete: 'F1' },
-        ];
-
-        const api = gridsManager.createGrid('grid-filter-removes-group-then-add', {
-            columnDefs: [{ field: 'country', rowGroup: true, hide: true }, { field: 'athlete' }],
-            autoGroupColumnDef: { headerName: 'Country' },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        await new GridRows(api, 'initial').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Ireland ag-Grid-AutoColumn:"Ireland"
-            │ └── LEAF id:1 country:"Ireland" athlete:"I1"
-            ├─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ └── LEAF id:2 country:"Italy" athlete:"T1"
-            └─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            · └── LEAF id:3 country:"France" athlete:"F1"
-        `);
-
-        api.setGridOption('quickFilterText', 'I1'); // shows only Ireland
-        await new GridRows(api, 'after filter Ireland only').check(`
-            ROOT id:ROOT_NODE_ID
-            └─┬ LEAF_GROUP id:row-group-country-Ireland ag-Grid-AutoColumn:"Ireland"
-            · └── LEAF id:1 country:"Ireland" athlete:"I1"
-        `);
-
-        api.setGridOption('quickFilterText', undefined);
-        applyTransactionChecked(api, { add: [{ id: '4', country: 'Spain', athlete: 'S1' }] });
-
-        await new GridRows(api, 'after add Spain').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Ireland ag-Grid-AutoColumn:"Ireland"
-            │ └── LEAF id:1 country:"Ireland" athlete:"I1"
-            ├─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ └── LEAF id:2 country:"Italy" athlete:"T1"
-            ├─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            │ └── LEAF id:3 country:"France" athlete:"F1"
-            └─┬ LEAF_GROUP id:row-group-country-Spain ag-Grid-AutoColumn:"Spain"
-            · └── LEAF id:4 country:"Spain" athlete:"S1"
-        `);
-    });
-
-    test('after removing a group, adding a new group appends at end (sentinel append)', async () => {
-        const rowData = [
-            { id: '1', country: 'Ireland', athlete: 'I1' },
-            { id: '2', country: 'Italy', athlete: 'It1' },
-            { id: '3', country: 'France', athlete: 'F1' },
-        ];
-
-        const api = gridsManager.createGrid('grid8', {
-            columnDefs: [{ field: 'country', rowGroup: true, hide: true }, { field: 'athlete' }],
-            autoGroupColumnDef: { headerName: 'Country' },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        await new GridRows(api, 'initial').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Ireland ag-Grid-AutoColumn:"Ireland"
-            │ └── LEAF id:1 country:"Ireland" athlete:"I1"
-            ├─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ └── LEAF id:2 country:"Italy" athlete:"It1"
-            └─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            · └── LEAF id:3 country:"France" athlete:"F1"
-        `);
-
-        applyTransactionChecked(api, { remove: [{ id: '2', country: 'Italy', athlete: 'It1' }] });
-
-        await new GridRows(api, 'after remove Italy').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Ireland ag-Grid-AutoColumn:"Ireland"
-            │ └── LEAF id:1 country:"Ireland" athlete:"I1"
-            └─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            · └── LEAF id:3 country:"France" athlete:"F1"
-        `);
-
-        applyTransactionChecked(api, { add: [{ id: '4', country: 'Spain', athlete: 'S1' }] });
-
-        await new GridRows(api, 'after add Spain').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Ireland ag-Grid-AutoColumn:"Ireland"
-            │ └── LEAF id:1 country:"Ireland" athlete:"I1"
-            ├─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            │ └── LEAF id:3 country:"France" athlete:"F1"
-            └─┬ LEAF_GROUP id:row-group-country-Spain ag-Grid-AutoColumn:"Spain"
-            · └── LEAF id:4 country:"Spain" athlete:"S1"
-        `);
-    });
-
-    test('apply group sort then clear: structural order is restored, not the prior sorted order', async () => {
-        const rowData = [
-            { id: '1', country: 'Audi', athlete: 'A' },
-            { id: '2', country: 'France', athlete: 'F' },
-            { id: '3', country: 'Italy', athlete: 'I' },
-        ];
-
-        const api = gridsManager.createGrid('grid-clear-sort', {
-            columnDefs: [{ field: 'country', rowGroup: true, hide: true, sortable: true }, { field: 'athlete' }],
-            autoGroupColumnDef: { headerName: 'Country' },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        await new GridRows(api, 'clear-sort: initial structural order').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            │ └── LEAF id:1 country:"Audi" athlete:"A"
-            ├─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            │ └── LEAF id:2 country:"France" athlete:"F"
-            └─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            · └── LEAF id:3 country:"Italy" athlete:"I"
-        `);
-
-        api.applyColumnState({ state: [{ colId: 'country', sort: 'desc' }] });
-        await new GridRows(api, 'clear-sort: country desc reorders groups').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ └── LEAF id:3 country:"Italy" athlete:"I"
-            ├─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            │ └── LEAF id:2 country:"France" athlete:"F"
-            └─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            · └── LEAF id:1 country:"Audi" athlete:"A"
-        `);
-
-        api.applyColumnState({ state: [{ colId: 'country', sort: null }] });
-        await new GridRows(api, 'clear-sort: structural order restored, not the prior desc order').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            │ └── LEAF id:1 country:"Audi" athlete:"A"
-            ├─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            │ └── LEAF id:2 country:"France" athlete:"F"
-            └─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            · └── LEAF id:3 country:"Italy" athlete:"I"
-        `);
-    });
-
-    test('single-sort: applying a leaf sort while a group sort is active clears the group sort and restores structural group order', async () => {
-        const rowData = [
-            { id: '1', country: 'Italy', athlete: 'It' },
-            { id: '2', country: 'France', athlete: 'Fr' },
-            { id: '3', country: 'Audi', athlete: 'Au' },
-        ];
-
-        const api = gridsManager.createGrid('grid-single-sort-clear-other', {
-            columnDefs: [
-                { field: 'country', rowGroup: true, hide: true, sortable: true },
-                { field: 'athlete', sortable: true },
-            ],
-            autoGroupColumnDef: { headerName: 'Country' },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        api.applyColumnState({ state: [{ colId: 'country', sort: 'asc' }] });
-        await new GridRows(api, 'single-sort: country asc').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            │ └── LEAF id:3 country:"Audi" athlete:"Au"
-            ├─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            │ └── LEAF id:2 country:"France" athlete:"Fr"
-            └─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            · └── LEAF id:1 country:"Italy" athlete:"It"
-        `);
-
-        api.applyColumnState({
-            state: [{ colId: 'athlete', sort: 'asc' }],
-            defaultState: { sort: null },
-        });
-        await new GridRows(api, 'single-sort: leaf sort active, group sort cleared — structural').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ └── LEAF id:1 country:"Italy" athlete:"It"
-            ├─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            │ └── LEAF id:2 country:"France" athlete:"Fr"
-            └─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            · └── LEAF id:3 country:"Audi" athlete:"Au"
-        `);
-    });
-
-    test('multi-level clear-sort: clearing one level restores structural order at that level only', async () => {
-        const rowData = [
-            { id: '1', country: 'Italy', year: 2021 },
-            { id: '2', country: 'Italy', year: 2020 },
-            { id: '3', country: 'France', year: 2019 },
-            { id: '4', country: 'France', year: 2022 },
-        ];
-
-        const api = gridsManager.createGrid('grid-multilevel-clear-sort', {
-            columnDefs: [
-                { field: 'country', rowGroup: true, hide: true, sortable: true },
-                { field: 'year', rowGroup: true, hide: true, sortable: true },
-            ],
-            autoGroupColumnDef: { headerName: 'Group' },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        api.applyColumnState({
-            state: [
-                { colId: 'country', sort: 'desc', sortIndex: 0 },
-                { colId: 'year', sort: 'asc', sortIndex: 1 },
-            ],
-        });
-        await new GridRows(api, 'multilevel: country desc + year asc').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ filler id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ ├─┬ LEAF_GROUP id:row-group-country-Italy-year-2020 ag-Grid-AutoColumn:2020
-            │ │ └── LEAF id:2 country:"Italy" year:2020
-            │ └─┬ LEAF_GROUP id:row-group-country-Italy-year-2021 ag-Grid-AutoColumn:2021
-            │ · └── LEAF id:1 country:"Italy" year:2021
-            └─┬ filler id:row-group-country-France ag-Grid-AutoColumn:"France"
-            · ├─┬ LEAF_GROUP id:row-group-country-France-year-2019 ag-Grid-AutoColumn:2019
-            · │ └── LEAF id:3 country:"France" year:2019
-            · └─┬ LEAF_GROUP id:row-group-country-France-year-2022 ag-Grid-AutoColumn:2022
-            · · └── LEAF id:4 country:"France" year:2022
-        `);
-
-        api.applyColumnState({ state: [{ colId: 'year', sort: null }] });
-        await new GridRows(api, 'multilevel: clear year — country desc kept, years structural').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ filler id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ ├─┬ LEAF_GROUP id:row-group-country-Italy-year-2021 ag-Grid-AutoColumn:2021
-            │ │ └── LEAF id:1 country:"Italy" year:2021
-            │ └─┬ LEAF_GROUP id:row-group-country-Italy-year-2020 ag-Grid-AutoColumn:2020
-            │ · └── LEAF id:2 country:"Italy" year:2020
-            └─┬ filler id:row-group-country-France ag-Grid-AutoColumn:"France"
-            · ├─┬ LEAF_GROUP id:row-group-country-France-year-2019 ag-Grid-AutoColumn:2019
-            · │ └── LEAF id:3 country:"France" year:2019
-            · └─┬ LEAF_GROUP id:row-group-country-France-year-2022 ag-Grid-AutoColumn:2022
-            · · └── LEAF id:4 country:"France" year:2022
-        `);
-
-        api.applyColumnState({ state: [{ colId: 'country', sort: null }] });
-        await new GridRows(api, 'multilevel: clear country too — fully structural').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ filler id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ ├─┬ LEAF_GROUP id:row-group-country-Italy-year-2021 ag-Grid-AutoColumn:2021
-            │ │ └── LEAF id:1 country:"Italy" year:2021
-            │ └─┬ LEAF_GROUP id:row-group-country-Italy-year-2020 ag-Grid-AutoColumn:2020
-            │ · └── LEAF id:2 country:"Italy" year:2020
-            └─┬ filler id:row-group-country-France ag-Grid-AutoColumn:"France"
-            · ├─┬ LEAF_GROUP id:row-group-country-France-year-2019 ag-Grid-AutoColumn:2019
-            · │ └── LEAF id:3 country:"France" year:2019
-            · └─┬ LEAF_GROUP id:row-group-country-France-year-2022 ag-Grid-AutoColumn:2022
-            · · └── LEAF id:4 country:"France" year:2022
-        `);
-    });
-
-    test('clear-sort after coincidental match: structural order persists after subsequent refresh', async () => {
-        const rowData = [
-            { id: '1', country: 'Audi', athlete: 'A' },
-            { id: '2', country: 'France', athlete: 'F' },
-            { id: '3', country: 'Italy', athlete: 'I' },
-        ];
-
-        const api = gridsManager.createGrid('grid-coincidental-match-clear', {
-            columnDefs: [{ field: 'country', rowGroup: true, hide: true, sortable: true }, { field: 'athlete' }],
-            autoGroupColumnDef: { headerName: 'Country' },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        // Sort country asc — produces [Audi, France, Italy] which equals structural order.
-        api.applyColumnState({ state: [{ colId: 'country', sort: 'asc' }] });
-        await new GridRows(api, 'coincidental: country asc matches structural').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            │ └── LEAF id:1 country:"Audi" athlete:"A"
-            ├─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            │ └── LEAF id:2 country:"France" athlete:"F"
-            └─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            · └── LEAF id:3 country:"Italy" athlete:"I"
-        `);
-
-        api.applyColumnState({ state: [{ colId: 'country', sort: null }] });
-        api.refreshClientSideRowModel('sort');
-        await new GridRows(api, 'coincidental: cleared and re-refreshed — structural').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Audi ag-Grid-AutoColumn:"Audi"
-            │ └── LEAF id:1 country:"Audi" athlete:"A"
-            ├─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            │ └── LEAF id:2 country:"France" athlete:"F"
-            └─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            · └── LEAF id:3 country:"Italy" athlete:"I"
-        `);
-    });
-
-    test('deltaSort + groupMaintainOrder: per-level isolation holds across transactions', async () => {
-        const rowData = [
-            { id: '1', country: 'Italy', year: 2021, athlete: 'Zed' },
-            { id: '2', country: 'Italy', year: 2021, athlete: 'Anna' },
-            { id: '3', country: 'France', year: 2019, athlete: 'Mark' },
-            { id: '4', country: 'France', year: 2019, athlete: 'Bob' },
-        ];
-
-        const api = createDeltaSortGrid('grid-delta-sort', {
-            columnDefs: [
-                { field: 'country', rowGroup: true, hide: true },
-                { field: 'year', rowGroup: true, hide: true },
-                { field: 'athlete' },
-            ],
-            autoGroupColumnDef: { headerName: 'Group' },
-            rowData,
-        });
-
-        api.applyColumnState({ state: [{ colId: 'athlete', sort: 'asc' }] });
-        await new GridRows(api, 'deltaSort: leaf rows sort, country/year groups stay structural').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ filler id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ └─┬ LEAF_GROUP id:row-group-country-Italy-year-2021 ag-Grid-AutoColumn:2021
-            │ · ├── LEAF id:2 country:"Italy" year:2021 athlete:"Anna"
-            │ · └── LEAF id:1 country:"Italy" year:2021 athlete:"Zed"
-            └─┬ filler id:row-group-country-France ag-Grid-AutoColumn:"France"
-            · └─┬ LEAF_GROUP id:row-group-country-France-year-2019 ag-Grid-AutoColumn:2019
-            · · ├── LEAF id:4 country:"France" year:2019 athlete:"Bob"
-            · · └── LEAF id:3 country:"France" year:2019 athlete:"Mark"
-        `);
-
-        // Transaction add: id=5 goes to Italy/2021. delta sort fires at the leaf group, structural
-        // baseline stays at country and year levels. Bart sorts between Anna and Zed.
-        applyTransactionChecked(api, { add: [{ id: '5', country: 'Italy', year: 2021, athlete: 'Bart' }] });
-        await new GridRows(api, 'deltaSort: after add — Bart placed in sorted leaf position').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ filler id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ └─┬ LEAF_GROUP id:row-group-country-Italy-year-2021 ag-Grid-AutoColumn:2021
-            │ · ├── LEAF id:2 country:"Italy" year:2021 athlete:"Anna"
-            │ · ├── LEAF id:5 country:"Italy" year:2021 athlete:"Bart"
-            │ · └── LEAF id:1 country:"Italy" year:2021 athlete:"Zed"
-            └─┬ filler id:row-group-country-France ag-Grid-AutoColumn:"France"
-            · └─┬ LEAF_GROUP id:row-group-country-France-year-2019 ag-Grid-AutoColumn:2019
-            · · ├── LEAF id:4 country:"France" year:2019 athlete:"Bob"
-            · · └── LEAF id:3 country:"France" year:2019 athlete:"Mark"
-        `);
-
-        // Update an athlete name to force a re-sort within the same leaf group; group order
-        // must remain unchanged.
-        applyTransactionChecked(api, {
-            update: [{ id: '2', country: 'Italy', year: 2021, athlete: 'Yves' }],
-        });
-        await new GridRows(api, 'deltaSort: after update — re-sorted leaf, groups still structural').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ filler id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ └─┬ LEAF_GROUP id:row-group-country-Italy-year-2021 ag-Grid-AutoColumn:2021
-            │ · ├── LEAF id:5 country:"Italy" year:2021 athlete:"Bart"
-            │ · ├── LEAF id:2 country:"Italy" year:2021 athlete:"Yves"
-            │ · └── LEAF id:1 country:"Italy" year:2021 athlete:"Zed"
-            └─┬ filler id:row-group-country-France ag-Grid-AutoColumn:"France"
-            · └─┬ LEAF_GROUP id:row-group-country-France-year-2019 ag-Grid-AutoColumn:2019
-            · · ├── LEAF id:4 country:"France" year:2019 athlete:"Bob"
-            · · └── LEAF id:3 country:"France" year:2019 athlete:"Mark"
-        `);
-    });
-
-    test('deltaSort + postSortRows-pinned group + transactions: per-level baseline integrity', async () => {
-        const rowData = [
-            { id: '1', country: 'Italy', athlete: 'Mark' },
-            { id: '2', country: 'France', athlete: 'Bob' },
-            { id: '3', country: 'Italy', athlete: 'Anna' },
-            { id: '4', country: 'France', athlete: 'Zed' },
-        ];
-
-        const api = createDeltaSortGrid('grid-delta-postsort', {
-            columnDefs: [{ field: 'country', rowGroup: true, hide: true }, { field: 'athlete' }],
-            autoGroupColumnDef: { headerName: 'Country' },
-            rowData,
-            postSortRows: (params: any) => {
-                const idx = params.nodes.findIndex((n: any) => n.key === 'Italy');
-                if (idx > 0) {
-                    const [pinned] = params.nodes.splice(idx, 1);
-                    params.nodes.unshift(pinned);
-                }
-            },
-        });
-
-        api.applyColumnState({ state: [{ colId: 'athlete', sort: 'asc' }] });
-
-        await new GridRows(api, 'delta+postSort: initial — Italy pinned, leaves sorted').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ ├── LEAF id:3 country:"Italy" athlete:"Anna"
-            │ └── LEAF id:1 country:"Italy" athlete:"Mark"
-            └─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            · ├── LEAF id:2 country:"France" athlete:"Bob"
-            · └── LEAF id:4 country:"France" athlete:"Zed"
-        `);
-
-        applyTransactionChecked(api, { add: [{ id: '5', country: 'France', athlete: 'Carl' }] });
-        await new GridRows(api, 'delta+postSort: add — Carl placed in sorted slot, Italy still pinned').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ ├── LEAF id:3 country:"Italy" athlete:"Anna"
-            │ └── LEAF id:1 country:"Italy" athlete:"Mark"
-            └─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            · ├── LEAF id:2 country:"France" athlete:"Bob"
-            · ├── LEAF id:5 country:"France" athlete:"Carl"
-            · └── LEAF id:4 country:"France" athlete:"Zed"
-        `);
-
-        applyTransactionChecked(api, { update: [{ id: '1', country: 'Italy', athlete: 'Yann' }] });
-        await new GridRows(api, 'delta+postSort: update reorders Italy leaves, Italy still pinned').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ ├── LEAF id:3 country:"Italy" athlete:"Anna"
-            │ └── LEAF id:1 country:"Italy" athlete:"Yann"
-            └─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            · ├── LEAF id:2 country:"France" athlete:"Bob"
-            · ├── LEAF id:5 country:"France" athlete:"Carl"
-            · └── LEAF id:4 country:"France" athlete:"Zed"
-        `);
-    });
-
-    test('deltaSort + postSortRows pinning at a sorted leaf level: full-sort fallback places new rows correctly', async () => {
-        const rowData = [
-            { id: '1', country: 'Italy', athlete: 'Mark' },
-            { id: '2', country: 'Italy', athlete: 'Anna' },
-            { id: '3', country: 'Italy', athlete: 'Carl' },
-            { id: '4', country: 'Italy', athlete: 'Bob' },
-            { id: '5', country: 'Italy', athlete: 'Zed' },
-            { id: '6', country: 'Italy', athlete: 'David' },
-        ];
-
-        const api = createDeltaSortGrid('grid-delta-postsort-sorted-leaf', {
-            columnDefs: [{ field: 'country', rowGroup: true, hide: true }, { field: 'athlete' }],
-            autoGroupColumnDef: { headerName: 'Country' },
-            rowData,
-            postSortRows: (params: any) => {
-                // Pin Zed to the front of any leaf array containing him.
-                const idx = params.nodes.findIndex((n: any) => n.data?.athlete === 'Zed');
-                if (idx > 0) {
-                    const [pinned] = params.nodes.splice(idx, 1);
-                    params.nodes.unshift(pinned);
-                }
-            },
-        });
-
-        api.applyColumnState({ state: [{ colId: 'athlete', sort: 'asc' }] });
-
-        await new GridRows(api, 'delta+postSort sorted leaf: initial — Zed pinned, rest asc').check(`
-            ROOT id:ROOT_NODE_ID
-            └─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            · ├── LEAF id:5 country:"Italy" athlete:"Zed"
-            · ├── LEAF id:2 country:"Italy" athlete:"Anna"
-            · ├── LEAF id:4 country:"Italy" athlete:"Bob"
-            · ├── LEAF id:3 country:"Italy" athlete:"Carl"
-            · ├── LEAF id:6 country:"Italy" athlete:"David"
-            · └── LEAF id:1 country:"Italy" athlete:"Mark"
-        `);
-
-        applyTransactionChecked(api, { add: [{ id: '7', country: 'Italy', athlete: 'Eric' }] });
-        await new GridRows(api, 'delta+postSort sorted leaf: add — Eric in correct sorted slot').check(`
-            ROOT id:ROOT_NODE_ID
-            └─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            · ├── LEAF id:5 country:"Italy" athlete:"Zed"
-            · ├── LEAF id:2 country:"Italy" athlete:"Anna"
-            · ├── LEAF id:4 country:"Italy" athlete:"Bob"
-            · ├── LEAF id:3 country:"Italy" athlete:"Carl"
-            · ├── LEAF id:6 country:"Italy" athlete:"David"
-            · ├── LEAF id:7 country:"Italy" athlete:"Eric"
-            · └── LEAF id:1 country:"Italy" athlete:"Mark"
-        `);
-
-        applyTransactionChecked(api, { update: [{ id: '1', country: 'Italy', athlete: 'Yann' }] });
-        await new GridRows(api, 'delta+postSort sorted leaf: update reorders, Zed still pinned').check(`
-            ROOT id:ROOT_NODE_ID
-            └─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            · ├── LEAF id:5 country:"Italy" athlete:"Zed"
-            · ├── LEAF id:2 country:"Italy" athlete:"Anna"
-            · ├── LEAF id:4 country:"Italy" athlete:"Bob"
-            · ├── LEAF id:3 country:"Italy" athlete:"Carl"
-            · ├── LEAF id:6 country:"Italy" athlete:"David"
-            · ├── LEAF id:7 country:"Italy" athlete:"Eric"
-            · └── LEAF id:1 country:"Italy" athlete:"Yann"
-        `);
-    });
-
-    test('treeData + groupMaintainOrder: per-level isolation does not apply, sort still runs', async () => {
-        const rowData = [
-            { id: '1', path: ['Audi'], name: 'Audi' },
-            { id: '2', path: ['Audi', 'A2'], name: 'A2' },
-            { id: '3', path: ['Audi', 'A1'], name: 'A1' },
-            { id: '4', path: ['BMW'], name: 'BMW' },
-            { id: '5', path: ['BMW', 'B1'], name: 'B1' },
-        ];
-
-        const api = gridsManager.createGrid('grid-tree-data-maintain', {
-            columnDefs: [{ field: 'name', sortable: true }],
-            treeData: true,
-            getDataPath: (data: any) => data.path,
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        await new GridRows(api, 'tree+maintain: initial structural').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ Audi GROUP id:1 ag-Grid-AutoColumn:"Audi" name:"Audi"
-            │ ├── A2 LEAF id:2 ag-Grid-AutoColumn:"A2" name:"A2"
-            │ └── A1 LEAF id:3 ag-Grid-AutoColumn:"A1" name:"A1"
-            └─┬ BMW GROUP id:4 ag-Grid-AutoColumn:"BMW" name:"BMW"
-            · └── B1 LEAF id:5 ag-Grid-AutoColumn:"B1" name:"B1"
-        `);
-
-        api.applyColumnState({ state: [{ colId: 'name', sort: 'asc' }] });
-        await new GridRows(api, 'tree+maintain: sort cascades — full reorder, no isolation').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ Audi GROUP id:1 ag-Grid-AutoColumn:"Audi" name:"Audi"
-            │ ├── A1 LEAF id:3 ag-Grid-AutoColumn:"A1" name:"A1"
-            │ └── A2 LEAF id:2 ag-Grid-AutoColumn:"A2" name:"A2"
-            └─┬ BMW GROUP id:4 ag-Grid-AutoColumn:"BMW" name:"BMW"
-            · └── B1 LEAF id:5 ag-Grid-AutoColumn:"B1" name:"B1"
-        `);
-
-        api.applyColumnState({ state: [{ colId: 'name', sort: null }] });
-        await new GridRows(api, 'tree+maintain: clear sort — structural order restored').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ Audi GROUP id:1 ag-Grid-AutoColumn:"Audi" name:"Audi"
-            │ ├── A2 LEAF id:2 ag-Grid-AutoColumn:"A2" name:"A2"
-            │ └── A1 LEAF id:3 ag-Grid-AutoColumn:"A1" name:"A1"
-            └─┬ BMW GROUP id:4 ag-Grid-AutoColumn:"BMW" name:"BMW"
-            · └── B1 LEAF id:5 ag-Grid-AutoColumn:"B1" name:"B1"
-        `);
-    });
-
-    test('deltaSort + filter cycle interleaved with transactions: leaves stay correctly sorted', async () => {
-        const rowData = [
-            { id: '1', country: 'Italy', athlete: 'Mark' },
-            { id: '2', country: 'France', athlete: 'Bob' },
-            { id: '3', country: 'Italy', athlete: 'Anna' },
-            { id: '4', country: 'France', athlete: 'Zed' },
-            { id: '5', country: 'Spain', athlete: 'Carlos' },
-        ];
-
-        const api = createDeltaSortGrid('grid-delta-filter-cycle', {
-            columnDefs: [{ field: 'country', rowGroup: true, hide: true }, { field: 'athlete' }],
-            autoGroupColumnDef: { headerName: 'Country' },
-            rowData,
-        });
-
-        api.applyColumnState({ state: [{ colId: 'athlete', sort: 'asc' }] });
-
-        await new GridRows(api, 'delta-filter: initial').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ ├── LEAF id:3 country:"Italy" athlete:"Anna"
-            │ └── LEAF id:1 country:"Italy" athlete:"Mark"
-            ├─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            │ ├── LEAF id:2 country:"France" athlete:"Bob"
-            │ └── LEAF id:4 country:"France" athlete:"Zed"
-            └─┬ LEAF_GROUP id:row-group-country-Spain ag-Grid-AutoColumn:"Spain"
-            · └── LEAF id:5 country:"Spain" athlete:"Carlos"
-        `);
-
-        api.setGridOption('quickFilterText', 'France');
-        await new GridRows(api, 'delta-filter: filtered to France').check(`
-            ROOT id:ROOT_NODE_ID
-            └─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            · ├── LEAF id:2 country:"France" athlete:"Bob"
-            · └── LEAF id:4 country:"France" athlete:"Zed"
-        `);
-
-        applyTransactionChecked(api, { add: [{ id: '6', country: 'Italy', athlete: 'Aaron' }] });
-        await new GridRows(api, 'delta-filter: add to hidden group while filtered').check(`
-            ROOT id:ROOT_NODE_ID
-            └─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            · ├── LEAF id:2 country:"France" athlete:"Bob"
-            · └── LEAF id:4 country:"France" athlete:"Zed"
-        `);
-
-        applyTransactionChecked(api, { update: [{ id: '5', country: 'Spain', athlete: 'Aldo' }] });
-
-        api.setGridOption('quickFilterText', undefined);
-        await new GridRows(api, 'delta-filter: clear filter — leaves still correctly sorted').check(`
-            ROOT id:ROOT_NODE_ID
-            ├─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            │ ├── LEAF id:6 country:"Italy" athlete:"Aaron"
-            │ ├── LEAF id:3 country:"Italy" athlete:"Anna"
-            │ └── LEAF id:1 country:"Italy" athlete:"Mark"
-            ├─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France"
-            │ ├── LEAF id:2 country:"France" athlete:"Bob"
-            │ └── LEAF id:4 country:"France" athlete:"Zed"
-            └─┬ LEAF_GROUP id:row-group-country-Spain ag-Grid-AutoColumn:"Spain"
-            · └── LEAF id:5 country:"Spain" athlete:"Aldo"
-        `);
-    });
-
-    // 6-row leaf group used by the deltaSort tests below. ≥5 rows is required to exercise the
-    // _doDeltaSort merge path (it falls back to full sort for unsortedRowsLen ≤ MIN_DELTA_SORT_ROWS=4).
-    const ITALY_SIX_ROWS = [
-        { id: '1', country: 'Italy', athlete: 'Mark' },
-        { id: '2', country: 'Italy', athlete: 'Anna' },
-        { id: '3', country: 'Italy', athlete: 'Carl' },
-        { id: '4', country: 'Italy', athlete: 'Bob' },
-        { id: '5', country: 'Italy', athlete: 'Zed' },
-        { id: '6', country: 'Italy', athlete: 'David' },
-    ] as const;
-
-    const ITALY_SIX_ROWS_ASC = `
-            ROOT id:ROOT_NODE_ID
-            └─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            · ├── LEAF id:2 country:"Italy" athlete:"Anna"
-            · ├── LEAF id:4 country:"Italy" athlete:"Bob"
-            · ├── LEAF id:3 country:"Italy" athlete:"Carl"
-            · ├── LEAF id:6 country:"Italy" athlete:"David"
-            · ├── LEAF id:1 country:"Italy" athlete:"Mark"
-            · └── LEAF id:5 country:"Italy" athlete:"Zed"
-        `;
-
-    // Generic helper for delta-sort + groupMaintainOrder grids — each test supplies its own
-    // columnDefs / rowData / extras. Shared options (animateRows, groupDefaultExpanded,
-    // groupMaintainOrder, deltaSort, getRowId) are the load-bearing setup for every delta-sort
-    // scenario in this file.
-    const createDeltaSortGrid = (gridName: string, overrides: Record<string, any>) =>
-        gridsManager.createGrid(gridName, {
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            deltaSort: true,
-            getRowId: (p: any) => p.data.id,
-            ...overrides,
-        });
-
-    // Italy/athlete shorthand — used by the 6-row ITALY_SIX_ROWS tests.
-    const createDeltaSortItalyGrid = (gridName: string) =>
-        createDeltaSortGrid(gridName, {
-            columnDefs: [{ field: 'country', rowGroup: true, hide: true }, { field: 'athlete' }],
-            autoGroupColumnDef: { headerName: 'Country' },
-            rowData: ITALY_SIX_ROWS.map((r) => ({ ...r })),
-        });
-
-    test('deltaSort + filter cycle on a 6-row leaf group: re-entered rows are correctly placed', async () => {
-        const api = createDeltaSortItalyGrid('grid-delta-large-leaf');
-        api.applyColumnState({ state: [{ colId: 'athlete', sort: 'asc' }] });
-        await new GridRows(api, 'delta-large: initial sorted').check(ITALY_SIX_ROWS_ASC);
-
-        // Filter to 'r' — matches Carl + Mark only (2 rows → full-sort fallback path).
-        api.setGridOption('quickFilterText', 'r');
-        await new GridRows(api, 'delta-large: filter to 2 rows').check(`
-            ROOT id:ROOT_NODE_ID
-            └─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            · ├── LEAF id:3 country:"Italy" athlete:"Carl"
-            · └── LEAF id:1 country:"Italy" athlete:"Mark"
-        `);
-
-        // Clear filter — re-entered rows are NOT in changedRowNodes; they must still land in
-        // the correct sorted positions on the merge path.
-        api.setGridOption('quickFilterText', undefined);
-        await new GridRows(api, 'delta-large: clear filter — all 6 rows correctly sorted').check(ITALY_SIX_ROWS_ASC);
-    });
-
-    test('deltaSort: changing sort direction triggers full sort, then transactions use the fresh baseline', async () => {
-        // Sort-option changes refresh WITHOUT a transaction → full-sort path rebuilds the
-        // baseline under the new options. The next transaction then engages delta sort with a
-        // baseline that matches the current sortOptionsForLevel.
-        const api = createDeltaSortItalyGrid('grid-delta-direction-flip');
-
-        api.applyColumnState({ state: [{ colId: 'athlete', sort: 'asc' }] });
-        await new GridRows(api, 'delta-direction: initial asc').check(ITALY_SIX_ROWS_ASC);
-
-        api.applyColumnState({ state: [{ colId: 'athlete', sort: 'desc' }] });
-        await new GridRows(api, 'delta-direction: flipped to desc').check(`
-            ROOT id:ROOT_NODE_ID
-            └─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            · ├── LEAF id:5 country:"Italy" athlete:"Zed"
-            · ├── LEAF id:1 country:"Italy" athlete:"Mark"
-            · ├── LEAF id:6 country:"Italy" athlete:"David"
-            · ├── LEAF id:3 country:"Italy" athlete:"Carl"
-            · ├── LEAF id:4 country:"Italy" athlete:"Bob"
-            · └── LEAF id:2 country:"Italy" athlete:"Anna"
-        `);
-
-        // Add a row — delta sort engages with the desc baseline. Eric must land between Mark
-        // and David in desc order (after Mark, before David).
-        applyTransactionChecked(api, { add: [{ id: '7', country: 'Italy', athlete: 'Eric' }] });
-        await new GridRows(api, 'delta-direction: add Eric — placed in desc slot').check(`
-            ROOT id:ROOT_NODE_ID
-            └─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            · ├── LEAF id:5 country:"Italy" athlete:"Zed"
-            · ├── LEAF id:1 country:"Italy" athlete:"Mark"
-            · ├── LEAF id:7 country:"Italy" athlete:"Eric"
-            · ├── LEAF id:6 country:"Italy" athlete:"David"
-            · ├── LEAF id:3 country:"Italy" athlete:"Carl"
-            · ├── LEAF id:4 country:"Italy" athlete:"Bob"
-            · └── LEAF id:2 country:"Italy" athlete:"Anna"
-        `);
-    });
-
-    test('deltaSort + leaf group fully filtered out, mutated while hidden, then re-shown: leaves still correctly sorted', async () => {
-        // Filter out every row in the only leaf group, mutate a hidden row via transaction, then
-        // clear the filter. ≥5 rows after re-show ensures the merge path runs in _doDeltaSort.
-        const api = createDeltaSortItalyGrid('grid-delta-fully-filtered');
-        api.applyColumnState({ state: [{ colId: 'athlete', sort: 'asc' }] });
-        await new GridRows(api, 'fully-filtered: initial sorted').check(ITALY_SIX_ROWS_ASC);
-
-        // Filter to a substring no row matches — Italy group becomes fully filtered out.
-        api.setGridOption('quickFilterText', 'NO-MATCH-TOKEN');
-        await new GridRows(api, 'fully-filtered: nothing visible').check(`
-            ROOT id:ROOT_NODE_ID
-        `);
-
-        // Mutate a hidden row's athlete: 'Anna' → 'Yvette' (was first asc, becomes near-last).
-        applyTransactionChecked(api, { update: [{ id: '2', country: 'Italy', athlete: 'Yvette' }] });
-
-        // Clear filter — all 6 rows reappear, sorted asc with Yvette in its new position.
-        api.setGridOption('quickFilterText', undefined);
-        await new GridRows(api, 'fully-filtered: re-shown — leaves sorted with mutation applied').check(`
-            ROOT id:ROOT_NODE_ID
-            └─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy"
-            · ├── LEAF id:4 country:"Italy" athlete:"Bob"
-            · ├── LEAF id:3 country:"Italy" athlete:"Carl"
-            · ├── LEAF id:6 country:"Italy" athlete:"David"
-            · ├── LEAF id:1 country:"Italy" athlete:"Mark"
-            · ├── LEAF id:2 country:"Italy" athlete:"Yvette"
-            · └── LEAF id:5 country:"Italy" athlete:"Zed"
-        `);
-    });
-
-    test('manual showRowGroup + own field: leaves reorder by the column own data', async () => {
-        const rowData = [
-            { id: '1', country: 'Italy', athlete: 'A1', label: 'M' },
-            { id: '2', country: 'Italy', athlete: 'A2', label: 'A' },
-            { id: '3', country: 'France', athlete: 'B1', label: 'Z' },
-        ];
-
-        const api = gridsManager.createGrid('grid-manual-showrowgroup-own-data', {
-            columnDefs: [
-                { field: 'country', rowGroup: true, hide: true },
-                { field: 'athlete' },
-                {
-                    colId: 'manualDisplay',
-                    headerName: 'Manual',
-                    showRowGroup: 'country',
-                    field: 'label',
-                    sortable: true,
-                    cellRenderer: 'agGroupCellRenderer',
-                    cellRendererParams: { suppressCount: true },
-                },
-            ],
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        api.applyColumnState({ state: [{ colId: 'manualDisplay', sort: 'asc' }] });
-
-        await new GridRows(api, 'manual showRowGroup + own field: leaves reorder by label').check(`
-            ROOT id:ROOT_NODE_ID manualDisplay:null
-            ├─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy" manualDisplay:"Italy"
-            │ ├── LEAF id:2 country:"Italy" athlete:"A2" manualDisplay:"A"
-            │ └── LEAF id:1 country:"Italy" athlete:"A1" manualDisplay:"M"
-            └─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France" manualDisplay:"France"
-            · └── LEAF id:3 country:"France" athlete:"B1" manualDisplay:"Z"
-        `);
-    });
-
-    test('uncoupled manual showRowGroup with custom comparator: groups reorder by the manual column comparator', async () => {
-        const rowData = [
-            { id: '1', country: 'Italy', athlete: 'A1', label: 'M' },
-            { id: '2', country: 'Italy', athlete: 'A2', label: 'A' },
-            { id: '3', country: 'France', athlete: 'B1', label: 'Z' },
-            { id: '4', country: 'BMW', athlete: 'C1', label: 'Q' },
-        ];
-
-        const api = gridsManager.createGrid('grid-manual-showrowgroup-own-data-uncoupled', {
-            columnDefs: [
-                { field: 'country', rowGroup: true, hide: true },
-                { field: 'athlete' },
-                {
-                    colId: 'manualDisplay',
-                    headerName: 'Manual',
-                    showRowGroup: 'country',
-                    field: 'label',
-                    sortable: true,
-                    comparator: (a: unknown, b: unknown) => {
-                        const aLen = a == null ? 0 : String(a).length;
-                        const bLen = b == null ? 0 : String(b).length;
-                        return aLen - bLen;
-                    },
-                    cellRenderer: 'agGroupCellRenderer',
-                    cellRendererParams: { suppressCount: true },
-                },
-            ],
-            autoGroupColumnDef: { comparator: () => 0 },
-            animateRows: false,
-            groupDefaultExpanded: -1,
-            groupMaintainOrder: true,
-            rowData,
-            getRowId: (p) => p.data.id,
-        });
-
-        api.applyColumnState({ state: [{ colId: 'manualDisplay', sort: 'asc' }] });
-
-        await new GridRows(api, 'manual showRowGroup sort: groups reorder by length').check(`
-            ROOT id:ROOT_NODE_ID manualDisplay:null
-            ├─┬ LEAF_GROUP id:row-group-country-BMW ag-Grid-AutoColumn:"BMW" manualDisplay:"BMW"
-            │ └── LEAF id:4 country:"BMW" athlete:"C1" manualDisplay:"Q"
-            ├─┬ LEAF_GROUP id:row-group-country-Italy ag-Grid-AutoColumn:"Italy" manualDisplay:"Italy"
-            │ ├── LEAF id:1 country:"Italy" athlete:"A1" manualDisplay:"M"
-            │ └── LEAF id:2 country:"Italy" athlete:"A2" manualDisplay:"A"
-            └─┬ LEAF_GROUP id:row-group-country-France ag-Grid-AutoColumn:"France" manualDisplay:"France"
-            · └── LEAF id:3 country:"France" athlete:"B1" manualDisplay:"Z"
-        `);
-    });
-});
